admiral: ADaM in R Asset Library

A toolbox for programming Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam).

See also

Author

Maintainer: Ben Straub ben.x.straub@gsk.com

Authors:

Stefan Bundfuss (ORCID)

Matt Bearham

Arianna Cascone (ORCID)

Kristin Dahnert

Jeffrey Dickinson

Ross Farrugia

Fanny Gautier

Edoardo Mancini (ORCID)

Gordon Miller

Lina Patil

Jim Rothstein (ORCID)

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]