Is an Argument a Character Scalar (String)?

Checks if an argument is a character scalar and (optionally) whether it matches one of the provided values.

assert_character_scalar(
  arg,
  values = NULL,
  case_sensitive = TRUE,
  optional = FALSE
)

Arguments

arg
A function argument to be checked

values
A character vector of valid values for arg

case_sensitive
Should the argument be handled case-sensitive? If set to FALSE, the argument is converted to lower case for checking the permitted values and returning the argument.

optional
Is the checked parameter optional? If set to FALSE and arg is NULL then an error is thrown

Value

The function throws an error if arg is not a character vector or if arg is a character vector but of length > 1 or if its value is not one of the values specified. Otherwise, the input is returned invisibly.

Examples

example_fun <- function(msg_type) {
  assert_character_scalar(msg_type, values = c("warning", "error"))
}

example_fun("warning")

try(example_fun("message"))

#> Error : `msg_type` must be one of 'warning' or 'error' but is 'message'

try(example_fun(TRUE))

#> Error : `msg_type` must be a character scalar but is `TRUE`

# handling parameters case-insensitive
example_fun2 <- function(msg_type) {
  msg_type <- assert_character_scalar(msg_type,
                                      values = c("warning", "error"),
                                      case_sensitive = FALSE)
 if (msg_type == "warning") {
   print("A warning was requested.")
 }
}

example_fun2("Warning")

#> [1] "A warning was requested."