Asserts parameters Argument and Converts to List of Expressions

The function asserts that the argument is a character vector or a list of expressions. If it is a character vector, it converts it to a list of symbols.

Usage

assert_parameters_argument(parameters, optional = TRUE)

Arguments

parameters

The argument to check

optional

Is the checked argument optional? If set to FALSE and parameters is NULL then an error is thrown.

Value

The parameters argument (converted to a list of symbol, if it is a character vector)