Create a basket_select object

Create a basket_select object

Usage

basket_select(name = NULL, id = NULL, scope = NULL, type)

Arguments

name

Name of the query used to select the definition of the query from the company database.

id

Identifier of the query used to select the definition of the query from the company database.

scope

Scope of the query used to select the definition of the query from the company database.

Permitted Values: "BROAD", "NARROW", NA_character_

type

The type argument expects a character scalar. It is passed to the company specific get_terms() function such that the function can determine which sort of basket is requested

Value

An object of class basket_select.

Details

Exactly one of name or id must be specified.

See also