Output a Dataset in a Vignette in the admiral Format

Output a dataset in a vignette with the pre-specified admiral format.

Usage

dataset_vignette(dataset, display_vars = NULL, filter = NULL)

Arguments

dataset

Dataset to output in the vignette

display_vars

Variables selected to demonstrate the outcome of the derivation

Permitted Values: list of variables

Default is NULL

If display_vars is not NULL, only the selected variables are visible in the vignette while the other variables are hidden. They can be made visible by clicking theChoose the columns to display button.

filter

Filter condition

The specified condition is applied to the dataset before it is displayed.

Permitted Values: a condition

Value

A HTML table

Examples

library(admiral)
library(DT)
library(dplyr)
library(admiraltest)
data("admiral_dm")

dataset_vignette(admiral_dm)

dataset_vignette(admiral_dm,
  display_vars = vars(USUBJID, RFSTDTC, DTHDTC),
  filter = ARMCD == "Pbo"
)