Derive a Disposition Date

Derive a disposition status date from the the relevant records in the disposition domain.

derive_disposition_dt(
  dataset,
  dataset_ds,
  new_var,
  dtc,
  filter_ds,
  date_imputation = NULL,
  subject_keys = vars(STUDYID, USUBJID)
)

Arguments

dataset
Input dataset

dataset_ds
Datasets containing the disposition information (e.g.: ds)

It must contain:

STUDYID, USUBJID,

The variable(s) specified in the dtc

The variables used in filter_ds.

new_var
Name of the disposition date variable

a variable name is expected

dtc
The character date used to derive/impute the disposition date

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss. If the year part is not recorded (missing date), no imputation is performed.

filter_ds
Filter condition for the disposition data.

Filter used to select the relevant disposition data. It is expected that the filter restricts dataset_ds such that there is at most one observation per patient. An error is issued otherwise.

Permitted Values: logical expression.

date_imputation
The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a

format with day and month specified as 'dd-mm': e.g. '15-06' for the 15th of June

or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month

Default is NULL

subject_keys
Variables to uniquely identify a subject

A list of quosures where the expressions are symbols as returned by vars() is expected.

Value

the input dataset with the disposition date (new_var) added

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data("dm")
data("ds")

dm %>%
  derive_disposition_dt(
    dataset_ds = ds,
    new_var = FRVDT,
    dtc = DSSTDTC,
    filter_ds = DSCAT == "OTHER EVENT" & DSDECOD == "FINAL RETRIEVAL VISIT"
  ) %>%
  select(STUDYID, USUBJID, FRVDT)

#> # A tibble: 306 x 3
#>    STUDYID      USUBJID     FRVDT     
#>    <chr>        <chr>       <date>    
#>  1 CDISCPILOT01 01-701-1015 NA        
#>  2 CDISCPILOT01 01-701-1023 2013-02-18
#>  3 CDISCPILOT01 01-701-1028 NA        
#>  4 CDISCPILOT01 01-701-1033 2014-09-15
#>  5 CDISCPILOT01 01-701-1034 NA        
#>  6 CDISCPILOT01 01-701-1047 2013-07-28
#>  7 CDISCPILOT01 01-701-1057 NA        
#>  8 CDISCPILOT01 01-701-1097 NA        
#>  9 CDISCPILOT01 01-701-1111 2013-02-22
#> 10 CDISCPILOT01 01-701-1115 2013-05-20
#> # … with 296 more rows