Derive a Disposition Reason at a Specific Timepoint

Derive a disposition reason from the the relevant records in the disposition domain.

derive_disposition_reason(
  dataset,
  dataset_ds,
  new_var,
  reason_var,
  new_var_spe = NULL,
  reason_var_spe = NULL,
  format_new_vars = format_reason_default,
  filter_ds
)

Arguments

dataset
Input dataset.

dataset_ds
Dataset containing the disposition information (e.g.: ds).

It must contains

STUDYID, USUBJID,

The variable(s) specified in the reason_var (and reason_var_spe, if required)

The variables used in filter_ds.

new_var
Name of the disposition reason variable.

A variable name is expected (e.g. DCSREAS).

reason_var
The variable used to derive the disposition reason

A variable name is expected (e.g. DSDECOD).

new_var_spe
Name of the disposition reason detail variable.

A variable name is expected (e.g. DCSREASP). If new_var_spe is specified, it is expected that reason_var_spe is also specified, otherwise an error is issued.

Default: NULL

reason_var_spe
The variable used to derive the disposition reason detail

A variable name is expected (e.g. DSTERM). If new_var_spe is specified, it is expected that reason_var_spe is also specified, otherwise an error is issued.

Default: NULL

format_new_vars
The function used to derive the reason(s)

This function is used to derive the disposition reason(s) and must follow the below conventions

If only the main reason for discontinuation needs to be derived (i.e. new_var_spe is NULL), the function must have at least one character vector argument, e.g. format_reason <- function(reason) and new_var will be derived as new_var = format_reason(reason_var) Typically, the content of the function would return reason_var or NA depending on the value (e.g. if_else ( reason != "COMPLETED" & !is.na(reason), reason, NA_character_)). DCSREAS = format_reason(DSDECOD) returns DCSREAS = DSDECOD when DSDECOD is not 'COMPLETED' nor NA, NA otherwise.

If both the main reason and the details needs to be derived (new_var_spe is specified) the function must have two character vectors argument, e.g. format_reason2 <- function(reason, reason_spe) and new_var will be derived as new_var = format_reason(reason_var), new_var_spe will be derived as new_var_spe = format_reason(reason_var, reason_var_spe), Typically, the content of the function would return reason_var_spe or NA depending on the reason_var value (e.g. if_else ( reason != "COMPLETED" & !is.na(reason), reason_spe, NA_character_)). DCSREASP = format_reason(DSDECOD, DSTERM) returns DCSREASP = DSTERM when DSDECOD is not 'COMPLETED' nor NA.

Default: format_reason_default defined as: format_reason_default <- function(reason, reason_spe = NULL) out <- if ( is.null(reason_spe) ) reason else reason_spe if_else ( reason != "COMPLETED" & !is.na(reason), out, NA_character_) format_reason_default(DSDECOD) returns DSDECOD when DSDECOD is not 'COMPLETED' nor NA. format_reason_default(DSDECOD, DSTERM) returns DSTERM when DSDECOD is not 'COMPLETED' nor NA.

filter_ds
Filter condition for the disposition data.

one observation per patient. An error is issued otherwise.

Permitted Values: logical expression.

Value

the input dataset with the disposition reason(s) (new_var and if required new_var_spe) added.

Details

This functions returns the main reason for discontinuation (e.g. DCSREAS or DCTREAS). The reason for discontinuation is derived based on reason_var (e.g. DSDECOD) and format_new_vars. If new_var_spe is not NULL, then the function will also returns the details associated with the reason for discontinuation (e.g. DCSREASP). The details associated with the reason for discontinuation are derived based on reason_var_spe (e.g. DSTERM), reason_var and format_new_vars.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
data("dm")
data("ds")

# Derive DCSREAS using the default format
derive_disposition_reason(
  dataset = dm,
  dataset_ds = ds,
  new_var = DCSREAS,
  reason_var = DSDECOD,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)

#> # A tibble: 306 x 26
#>    STUDYID DOMAIN USUBJID SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>   <chr>  <chr>   <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCP… DM     01-701… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCP… DM     01-701… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCP… DM     01-701… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCP… DM     01-701… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCP… DM     01-701… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCP… DM     01-701… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCP… DM     01-701… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCP… DM     01-701… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCP… DM     01-701… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCP… DM     01-701… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 17 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>, DCSREAS <chr>

# Derive DCSREAS and DCSREASP using a study-specific format
format_dcsreas <- function(x, y = NULL) {
  out <- if (is.null(y)) x else y
  case_when(
    !(x %in% c("COMPLETED", "SCREEN FAILURE")) & !is.na(x) ~ out,
    TRUE ~ NA_character_
  )
}
derive_disposition_reason(
  dataset = dm,
  dataset_ds = ds,
  new_var = DCSREAS,
  reason_var = DSDECOD,
  new_var_spe = DCSREASP,
  reason_var_spe = DSTERM,
  format_new_vars = format_dcsreas,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)

#> # A tibble: 306 x 27
#>    STUDYID DOMAIN USUBJID SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>   <chr>  <chr>   <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCP… DM     01-701… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCP… DM     01-701… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCP… DM     01-701… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCP… DM     01-701… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCP… DM     01-701… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCP… DM     01-701… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCP… DM     01-701… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCP… DM     01-701… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCP… DM     01-701… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCP… DM     01-701… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 18 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>, DCSREAS <chr>, DCSREASP <chr>