Adds a variable flagging the first or last observation within each by group

Adds a variable flagging the first or last observation within each by group

derive_extreme_flag(dataset, new_var, by_vars, order, mode = "last",
  flag_filter, check_type = "warning")

Arguments

dataset
Input dataset

The variables specified by the by_vars parameter are expected.

new_var
Variable to add

The specified variable is added to the output dataset. It is set to "Y" for the first or last observation (depending on the mode) of each by group.

Permitted Values: list of name-value pairs

by_vars
Grouping variables

Permitted Values: list of variables

order
Sort order

The first or last observation is determined with respect to the specified order.

Permitted Values: list of variables or functions of variables

mode
Flag mode

Determines of the first or last observation is flagged.

Default: "last"

Permitted Values: "first", "last"

flag_filter
Filter for flag data

Only observations fulfilling the specified condition are taken into account for flagging. If the parameter is not specified, all observations are considered.

Permitted Values: a condition

check_type
Check uniqueness?

If "warning" or "error" is specified, the specified message is issued if the observations of the input dataset are not unique with respect to the by variables and the order.

Default: "warning"

Permitted Values: "none", "warning", "error"

Value

The input dataset with the new flag variable added

Details

For each group (with respect to the variables specified for the by_vars parameter) the first or last observation (with respect to the order specified for the order parameter and the flag mode) is included in the output dataset.

Examples

library(dplyr, warn.conflicts = FALSE)
data("vs")

# flag last value for each patient, test, and visit, baseline observations are ignored
derive_extreme_flag(vs,
                    new_var = LASTFL,
                    by_vars = rlang::exprs(USUBJID, VSTESTCD, VISIT),
                    order = rlang::exprs(VSTPTNUM),
                    flag_filter = rlang::expr(VISIT != "BASELINE")) %>%
  arrange(USUBJID, VSTESTCD, VISITNUM, VSTPTNUM) %>%
  select(USUBJID, VSTESTCD, VISIT, VSTPTNUM, VSSTRESN, LASTFL)

#> # A tibble: 29,643 x 6
#>    USUBJID     VSTESTCD VISIT               VSTPTNUM VSSTRESN LASTFL
#>    <chr>       <chr>    <chr>                  <dbl>    <dbl> <chr> 
#>  1 01-701-1015 DIABP    SCREENING 1              815       64 <NA>  
#>  2 01-701-1015 DIABP    SCREENING 1              816       83 <NA>  
#>  3 01-701-1015 DIABP    SCREENING 1              817       57 Y     
#>  4 01-701-1015 DIABP    SCREENING 2              815       68 <NA>  
#>  5 01-701-1015 DIABP    SCREENING 2              816       59 <NA>  
#>  6 01-701-1015 DIABP    SCREENING 2              817       71 Y     
#>  7 01-701-1015 DIABP    BASELINE                 815       56 <NA>  
#>  8 01-701-1015 DIABP    BASELINE                 816       51 <NA>  
#>  9 01-701-1015 DIABP    BASELINE                 817       61 <NA>  
#> 10 01-701-1015 DIABP    AMBUL ECG PLACEMENT      815       67 <NA>  
#> # … with 29,633 more rows

# Baseline (ABLFL) examples:

input <- tibble::tribble(
~STUDYID, ~USUBJID, ~PARAMCD,  ~AVISIT,    ~ADT,                 ~AVAL, ~DTYPE,
"TEST01", "PAT01",  "PARAM01", "BASELINE", as.Date("2021-04-27"), 15.0, NA,
"TEST01", "PAT01",  "PARAM01", "BASELINE", as.Date("2021-04-25"), 14.0, NA,
"TEST01", "PAT01",  "PARAM01", "BASELINE", as.Date("2021-04-23"), 15.0, "AVERAGE",
"TEST01", "PAT01",  "PARAM01", "WEEK 1",   as.Date("2021-04-27"), 10.0, "AVERAGE",
"TEST01", "PAT01",  "PARAM01", "WEEK 2",   as.Date("2021-04-30"), 12.0, NA,

"TEST01", "PAT02",  "PARAM01", "SCREENING",as.Date("2021-04-27"), 15.0, "AVERAGE",
"TEST01", "PAT02",  "PARAM01", "BASELINE", as.Date("2021-04-25"), 14.0, "AVERAGE",
"TEST01", "PAT02",  "PARAM01", "BASELINE", as.Date("2021-04-23"), 15.0, "AVERAGE",
"TEST01", "PAT02",  "PARAM01", "WEEK 1",   as.Date("2021-04-27"), 10.0, "AVERAGE",
"TEST01", "PAT02",  "PARAM01", "WEEK 2",   as.Date("2021-04-30"), 12.0, "AVERAGE",

"TEST01", "PAT01",  "PARAM02", "SCREENING",as.Date("2021-04-27"), 15.0, "AVERAGE",
"TEST01", "PAT01",  "PARAM02", "SCREENING",as.Date("2021-04-25"), 14.0, "AVERAGE",
"TEST01", "PAT01",  "PARAM02", "SCREENING",as.Date("2021-04-23"), 15.0, NA,
"TEST01", "PAT01",  "PARAM02", "BASELINE", as.Date("2021-04-27"), 10.0, "AVERAGE",
"TEST01", "PAT01",  "PARAM02", "WEEK 2",   as.Date("2021-04-30"), 12.0, NA,

"TEST01", "PAT02",  "PARAM02", "SCREENING",as.Date("2021-04-27"), 15.0, NA,
"TEST01", "PAT02",  "PARAM02", "BASELINE", as.Date("2021-04-25"), 14.0, NA,
"TEST01", "PAT02",  "PARAM02", "WEEK 1",   as.Date("2021-04-23"), 15.0, NA,
"TEST01", "PAT02",  "PARAM02", "WEEK 1",   as.Date("2021-04-27"), 10.0, NA,
"TEST01", "PAT02",  "PARAM02", "BASELINE", as.Date("2021-04-30"), 12.0, NA
)

# Last observation
derive_extreme_flag(
  input,
  new_var = ABLFL,
  by_vars = exprs(USUBJID, PARAMCD),
  order = exprs(ADT),
  flag_filter = expr(AVISIT == "BASELINE")
)

#> # A tibble: 20 x 8
#>    STUDYID USUBJID PARAMCD AVISIT    ADT         AVAL DTYPE   ABLFL
#>    <chr>   <chr>   <chr>   <chr>     <date>     <dbl> <chr>   <chr>
#>  1 TEST01  PAT01   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  2 TEST01  PAT01   PARAM01 BASELINE  2021-04-25    14 <NA>    <NA> 
#>  3 TEST01  PAT01   PARAM01 BASELINE  2021-04-27    15 <NA>    Y    
#>  4 TEST01  PAT01   PARAM02 BASELINE  2021-04-27    10 AVERAGE Y    
#>  5 TEST01  PAT02   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  6 TEST01  PAT02   PARAM01 BASELINE  2021-04-25    14 AVERAGE Y    
#>  7 TEST01  PAT02   PARAM02 BASELINE  2021-04-25    14 <NA>    <NA> 
#>  8 TEST01  PAT02   PARAM02 BASELINE  2021-04-30    12 <NA>    Y    
#>  9 TEST01  PAT01   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 10 TEST01  PAT01   PARAM01 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 11 TEST01  PAT02   PARAM01 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 12 TEST01  PAT02   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 13 TEST01  PAT02   PARAM01 WEEK 2    2021-04-30    12 AVERAGE <NA> 
#> 14 TEST01  PAT01   PARAM02 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 15 TEST01  PAT01   PARAM02 SCREENING 2021-04-25    14 AVERAGE <NA> 
#> 16 TEST01  PAT01   PARAM02 SCREENING 2021-04-23    15 <NA>    <NA> 
#> 17 TEST01  PAT01   PARAM02 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 18 TEST01  PAT02   PARAM02 SCREENING 2021-04-27    15 <NA>    <NA> 
#> 19 TEST01  PAT02   PARAM02 WEEK 1    2021-04-23    15 <NA>    <NA> 
#> 20 TEST01  PAT02   PARAM02 WEEK 1    2021-04-27    10 <NA>    <NA> 

# Worst observation - Direction = High
derive_extreme_flag(
  input,
  new_var = ABLFL,
  by_vars = exprs(USUBJID, PARAMCD),
  order = exprs(AVAL, ADT),
  flag_filter = expr(AVISIT == "BASELINE")
)

#> # A tibble: 20 x 8
#>    STUDYID USUBJID PARAMCD AVISIT    ADT         AVAL DTYPE   ABLFL
#>    <chr>   <chr>   <chr>   <chr>     <date>     <dbl> <chr>   <chr>
#>  1 TEST01  PAT01   PARAM01 BASELINE  2021-04-25    14 <NA>    <NA> 
#>  2 TEST01  PAT01   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  3 TEST01  PAT01   PARAM01 BASELINE  2021-04-27    15 <NA>    Y    
#>  4 TEST01  PAT01   PARAM02 BASELINE  2021-04-27    10 AVERAGE Y    
#>  5 TEST01  PAT02   PARAM01 BASELINE  2021-04-25    14 AVERAGE <NA> 
#>  6 TEST01  PAT02   PARAM01 BASELINE  2021-04-23    15 AVERAGE Y    
#>  7 TEST01  PAT02   PARAM02 BASELINE  2021-04-30    12 <NA>    <NA> 
#>  8 TEST01  PAT02   PARAM02 BASELINE  2021-04-25    14 <NA>    Y    
#>  9 TEST01  PAT01   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 10 TEST01  PAT01   PARAM01 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 11 TEST01  PAT02   PARAM01 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 12 TEST01  PAT02   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 13 TEST01  PAT02   PARAM01 WEEK 2    2021-04-30    12 AVERAGE <NA> 
#> 14 TEST01  PAT01   PARAM02 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 15 TEST01  PAT01   PARAM02 SCREENING 2021-04-25    14 AVERAGE <NA> 
#> 16 TEST01  PAT01   PARAM02 SCREENING 2021-04-23    15 <NA>    <NA> 
#> 17 TEST01  PAT01   PARAM02 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 18 TEST01  PAT02   PARAM02 SCREENING 2021-04-27    15 <NA>    <NA> 
#> 19 TEST01  PAT02   PARAM02 WEEK 1    2021-04-23    15 <NA>    <NA> 
#> 20 TEST01  PAT02   PARAM02 WEEK 1    2021-04-27    10 <NA>    <NA> 

# Worst observation - Direction = Lo
derive_extreme_flag(
  input,
  new_var = ABLFL,
  by_vars = exprs(USUBJID, PARAMCD),
  order = exprs(desc(AVAL), ADT),
  flag_filter = expr(AVISIT == "BASELINE")
)

#> # A tibble: 20 x 8
#>    STUDYID USUBJID PARAMCD AVISIT    ADT         AVAL DTYPE   ABLFL
#>    <chr>   <chr>   <chr>   <chr>     <date>     <dbl> <chr>   <chr>
#>  1 TEST01  PAT01   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  2 TEST01  PAT01   PARAM01 BASELINE  2021-04-27    15 <NA>    <NA> 
#>  3 TEST01  PAT01   PARAM01 BASELINE  2021-04-25    14 <NA>    Y    
#>  4 TEST01  PAT01   PARAM02 BASELINE  2021-04-27    10 AVERAGE Y    
#>  5 TEST01  PAT02   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  6 TEST01  PAT02   PARAM01 BASELINE  2021-04-25    14 AVERAGE Y    
#>  7 TEST01  PAT02   PARAM02 BASELINE  2021-04-25    14 <NA>    <NA> 
#>  8 TEST01  PAT02   PARAM02 BASELINE  2021-04-30    12 <NA>    Y    
#>  9 TEST01  PAT01   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 10 TEST01  PAT01   PARAM01 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 11 TEST01  PAT02   PARAM01 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 12 TEST01  PAT02   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 13 TEST01  PAT02   PARAM01 WEEK 2    2021-04-30    12 AVERAGE <NA> 
#> 14 TEST01  PAT01   PARAM02 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 15 TEST01  PAT01   PARAM02 SCREENING 2021-04-25    14 AVERAGE <NA> 
#> 16 TEST01  PAT01   PARAM02 SCREENING 2021-04-23    15 <NA>    <NA> 
#> 17 TEST01  PAT01   PARAM02 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 18 TEST01  PAT02   PARAM02 SCREENING 2021-04-27    15 <NA>    <NA> 
#> 19 TEST01  PAT02   PARAM02 WEEK 1    2021-04-23    15 <NA>    <NA> 
#> 20 TEST01  PAT02   PARAM02 WEEK 1    2021-04-27    10 <NA>    <NA> 

# Average observation
derive_extreme_flag(
  input,
  new_var = ABLFL,
  by_vars = exprs(USUBJID, PARAMCD),
  order = exprs(ADT, desc(AVAL)),
  flag_filter = expr(AVISIT == "BASELINE" & DTYPE == "AVERAGE")
)

#> # A tibble: 20 x 8
#>    STUDYID USUBJID PARAMCD AVISIT    ADT         AVAL DTYPE   ABLFL
#>    <chr>   <chr>   <chr>   <chr>     <date>     <dbl> <chr>   <chr>
#>  1 TEST01  PAT01   PARAM01 BASELINE  2021-04-23    15 AVERAGE Y    
#>  2 TEST01  PAT01   PARAM02 BASELINE  2021-04-27    10 AVERAGE Y    
#>  3 TEST01  PAT02   PARAM01 BASELINE  2021-04-23    15 AVERAGE <NA> 
#>  4 TEST01  PAT02   PARAM01 BASELINE  2021-04-25    14 AVERAGE Y    
#>  5 TEST01  PAT01   PARAM01 BASELINE  2021-04-27    15 <NA>    <NA> 
#>  6 TEST01  PAT01   PARAM01 BASELINE  2021-04-25    14 <NA>    <NA> 
#>  7 TEST01  PAT01   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#>  8 TEST01  PAT01   PARAM01 WEEK 2    2021-04-30    12 <NA>    <NA> 
#>  9 TEST01  PAT02   PARAM01 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 10 TEST01  PAT02   PARAM01 WEEK 1    2021-04-27    10 AVERAGE <NA> 
#> 11 TEST01  PAT02   PARAM01 WEEK 2    2021-04-30    12 AVERAGE <NA> 
#> 12 TEST01  PAT01   PARAM02 SCREENING 2021-04-27    15 AVERAGE <NA> 
#> 13 TEST01  PAT01   PARAM02 SCREENING 2021-04-25    14 AVERAGE <NA> 
#> 14 TEST01  PAT01   PARAM02 SCREENING 2021-04-23    15 <NA>    <NA> 
#> 15 TEST01  PAT01   PARAM02 WEEK 2    2021-04-30    12 <NA>    <NA> 
#> 16 TEST01  PAT02   PARAM02 SCREENING 2021-04-27    15 <NA>    <NA> 
#> 17 TEST01  PAT02   PARAM02 BASELINE  2021-04-25    14 <NA>    <NA> 
#> 18 TEST01  PAT02   PARAM02 WEEK 1    2021-04-23    15 <NA>    <NA> 
#> 19 TEST01  PAT02   PARAM02 WEEK 1    2021-04-27    10 <NA>    <NA> 
#> 20 TEST01  PAT02   PARAM02 BASELINE  2021-04-30    12 <NA>    <NA>