Derive Last Dose Date(-time)

Deprecated, please use derive_var_last_dose_date() instead.

Usage

derive_last_dose(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  by_vars = vars(STUDYID, USUBJID),
  dose_start,
  dose_end,
  analysis_date,
  dataset_seq_var,
  new_var,
  output_datetime = TRUE,
  check_dates_only = FALSE,
  traceability_vars = NULL
)

Arguments

dataset

Input dataset.

dataset_ex

Input EX dataset.

filter_ex

Filtering condition applied to EX dataset. For example, it can be used to filter for valid dose. Defaults to NULL.

by_vars

Variables to join by (created by dplyr::vars).

dose_start

The dose start date variable.

dose_end

The dose end date variable.

analysis_date

The analysis date variable.

dataset_seq_var

The sequence variable (this together with by_vars creates the keys of dataset).

new_var

The output variable.

output_datetime

Logical. Should only date or date-time variable be returned? Defaults to TRUE (i.e. date-time variable).

check_dates_only

Logical. An assumption that start and end dates of treatment match is checked. By default (FALSE), the date as well as the time component is checked. If set to TRUE, then only the date component of those variables is checked.

traceability_vars

A named list returned by vars() listing the traceability variables, e.g. vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ). The left-hand side (names of the list elements) gives the names of the traceability variables in the returned dataset. The right-hand side (values of the list elements) gives the values of the traceability variables in the returned dataset. These can be either strings or symbols referring to existing variables.

Value

Input dataset with additional column new_var.

Details

All date (date-time) variables can be characters in standard ISO format or of date / date-time class. For ISO format, see impute_dtc - parameter dtc for further details. Date-time imputations are done as follows:

dose_end: no date imputation, time imputation to 00:00:00 if time is missing.

analysis_date: no date imputation, time imputation to 23:59:59 if time is missing.

The last dose date is derived as follows:

The dataset_ex is filtered using filter_ex, if provided. This is useful for, for example, filtering for valid dose only.

The datasets dataset and dataset_ex are joined using by_vars.

The last dose date is derived: the last dose date is the maximum date where dose_end is lower to or equal to analysis_date, subject to both date values are non-NA. The last dose date is derived per by_vars and dataset_seq_var.

The last dose date is appended to the dataset and returned to the user.

Furthermore, the following assumption is checked: start and end dates (datetimes) need to match. Use check_dates_only to control whether only dates or whole date-times need to be equal.

Author

Ondrej Slama

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data(admiral_ae)
data(ex_single)

admiral_ae %>%
  head(100) %>%
  derive_last_dose(
    head(ex_single, 100),
    filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
      nchar(EXENDTC) >= 10,
    dose_start = EXSTDTC,
    dose_end = EXENDTC,
    analysis_date = AESTDTC,
    dataset_seq_var = AESEQ,
    new_var = LDOSEDTM,
    output_datetime = TRUE,
    check_dates_only = FALSE
  ) %>%
  select(STUDYID, USUBJID, AESEQ, AESTDTC, LDOSEDTM)
#> Warning: `derive_last_dose()` was deprecated in admiral 0.6.0.
#> Please use `derive_var_last_dose_date()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> # A tibble: 100 x 5
#>    STUDYID      USUBJID     AESEQ AESTDTC    LDOSEDTM           
#>    <chr>        <chr>       <dbl> <chr>      <dttm>             
#>  1 CDISCPILOT01 01-701-1015     1 2014-01-03 NA                 
#>  2 CDISCPILOT01 01-701-1015     2 2014-01-03 NA                 
#>  3 CDISCPILOT01 01-701-1015     3 2014-01-09 NA                 
#>  4 CDISCPILOT01 01-701-1023     3 2012-08-26 2012-08-11 00:00:00
#>  5 CDISCPILOT01 01-701-1023     1 2012-08-07 2012-08-07 00:00:00
#>  6 CDISCPILOT01 01-701-1023     2 2012-08-07 2012-08-07 00:00:00
#>  7 CDISCPILOT01 01-701-1023     4 2012-08-07 2012-08-07 00:00:00
#>  8 CDISCPILOT01 01-701-1028     1 2013-07-21 NA                 
#>  9 CDISCPILOT01 01-701-1028     2 2013-08-08 NA                 
#> 10 CDISCPILOT01 01-701-1034     1 2014-08-27 NA                 
#> # … with 90 more rows

# or with traceability variables
admiral_ae %>%
  head(100) %>%
  derive_last_dose(
    head(ex_single, 100),
    filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
      nchar(EXENDTC) >= 10,
    dose_start = EXSTDTC,
    dose_end = EXENDTC,
    analysis_date = AESTDTC,
    dataset_seq_var = AESEQ,
    new_var = LDOSEDTM,
    output_datetime = TRUE,
    check_dates_only = FALSE,
    traceability_vars = dplyr::vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ, LDOSEVAR = "EXSTDTC")
  ) %>%
  select(STUDYID, USUBJID, AESEQ, AESTDTC, LDOSEDTM, LDOSEDOM, LDOSESEQ, LDOSEVAR)
#> # A tibble: 100 x 8
#>    STUDYID  USUBJID AESEQ AESTDTC LDOSEDTM            LDOSEDOM LDOSESEQ LDOSEVAR
#>    <chr>    <chr>   <dbl> <chr>   <dttm>              <chr>       <int> <chr>   
#>  1 CDISCPI… 01-701…     1 2014-0… NA                  NA             NA NA      
#>  2 CDISCPI… 01-701…     2 2014-0… NA                  NA             NA NA      
#>  3 CDISCPI… 01-701…     3 2014-0… NA                  NA             NA NA      
#>  4 CDISCPI… 01-701…     3 2012-0… 2012-08-11 00:00:00 EX              7 EXSTDTC 
#>  5 CDISCPI… 01-701…     1 2012-0… 2012-08-07 00:00:00 EX              3 EXSTDTC 
#>  6 CDISCPI… 01-701…     2 2012-0… 2012-08-07 00:00:00 EX              3 EXSTDTC 
#>  7 CDISCPI… 01-701…     4 2012-0… 2012-08-07 00:00:00 EX              3 EXSTDTC 
#>  8 CDISCPI… 01-701…     1 2013-0… NA                  NA             NA NA      
#>  9 CDISCPI… 01-701…     2 2013-0… NA                  NA             NA NA      
#> 10 CDISCPI… 01-701…     1 2014-0… NA                  NA             NA NA      
#> # … with 90 more rows