Merge variables from a second dataset

Merges variables from a second dataset. The observations to merge can be selected by a condition and/or by selecting the first observation in each by group.

derive_merged_vars(dataset, dataset_add, filter_add = NULL, new_vars,
  by_vars = exprs(USUBJID), filter_order = NULL,
  filter_mode = "last")

Arguments

dataset
Input dataset

The variables specified by the by_vars parameter are expected.

dataset_add
Dataset to add

The variables specified by the by_vars, the filter_add, and the filter_first_order parameter are expected.

filter_add
Filter condition for dataset to add

Only observations of the add dataset which fulfill the specified condition are used for merging.

Permitted Values: logical expression

new_vars
Variable to add

The specified variables are created in the add dataset and then merge to the input dataset.

Permitted Values: list of name-value pairs

by_vars
Grouping variables

Default: exprs(USUBJID)

Permitted Values: list of variables

filter_order
Sort order

If the parameter is specified, the add dataset is ordered by the specified order and only the first or last observation (depending on the filter mode) in each by group is used for merging.

Permitted Values: list of variables or functions of variables

filter_mode
Filter mode

If the filter_order parameter is specified, the filter mode determines if the first or last observation of each by group is selected.

Default: "last"

Permitted Values: "first", "last"

Value

The input dataset with variables from the add dataset

Details

For each group (with respect to the variables specified for the by_vars parameter) the first observation (with respect to the order specified for the order parameter) is included in the output dataset.

Examples

library(dplyr)
library(magrittr)

data("ex")
data("dm")

# adding first treatment start date for each patient
derive_merged_vars(dm,
                   dataset_add = ex,
                   filter_add = exprs(EXDOSE > 0 | str_detect(EXTRT, 'PLACEBO')),
                   new_vars = exprs(TRTSDT := ymd(EXSTDTC)),
                   filter_order = exprs(EXSTDTC, EXSEQ),
                   filter_mode = "first") %>%
  select(USUBJID, TRTSDT)

#> # A tibble: 306 x 2
#>    USUBJID     TRTSDT    
#>    <chr>       <date>    
#>  1 01-701-1015 2014-01-02
#>  2 01-701-1023 2012-08-05
#>  3 01-701-1028 2013-07-19
#>  4 01-701-1033 2014-03-18
#>  5 01-701-1034 2014-07-01
#>  6 01-701-1047 2013-02-12
#>  7 01-701-1057 NA        
#>  8 01-701-1097 2014-01-01
#>  9 01-701-1111 2012-09-07
#> 10 01-701-1115 2012-11-30
#> # … with 296 more rows