Adds a Variable Numbering the Observations Within Each By Group

Adds a variable numbering the observations within each by group

derive_obs_number(
  dataset,
  by_vars = NULL,
  order = NULL,
  new_var = ASEQ,
  check_type = "none"
)

Arguments

dataset
Input dataset

The variables specified by the order and the by_vars parameter are expected.

by_vars
Grouping variables

Permitted Values: list of variables

order
Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables

new_var
Name of variable to create

The new variable is set to the observation number for each by group. The numbering starts with 1.

Default: ASEQ

check_type
Check uniqueness?

If "warning" or "error" is specified, the specified message is issued if the observations of the input dataset are not unique with respect to the by variables and the order.

Default: "none"

Permitted Values: "none", "warning", "error"

Value

A dataset containing all observations and variables of the input dataset and additionally the variable specified by the new_var parameter.

Details

For each group (with respect to the variables specified for the by_vars parameter) the first or last observation (with respect to the order specified for the order parameter and the mode specified for the mode parameter) is included in the output dataset.

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data("vs")

vs %>%
  select(USUBJID, VSTESTCD, VISITNUM, VSTPTNUM) %>%
  filter(VSTESTCD %in% c("HEIGHT", "WEIGHT")) %>%
  derive_obs_number(
    by_vars = vars(USUBJID, VSTESTCD),
    order = vars(VISITNUM, VSTPTNUM)
  )

#> # A tibble: 2,304 x 5
#>    USUBJID     VSTESTCD VISITNUM VSTPTNUM  ASEQ
#>    <chr>       <chr>       <dbl>    <dbl> <int>
#>  1 01-701-1015 HEIGHT          1       NA     1
#>  2 01-701-1015 WEIGHT          1       NA     1
#>  3 01-701-1015 WEIGHT          3       NA     2
#>  4 01-701-1015 WEIGHT          4       NA     3
#>  5 01-701-1015 WEIGHT          5       NA     4
#>  6 01-701-1015 WEIGHT          7       NA     5
#>  7 01-701-1015 WEIGHT          8       NA     6
#>  8 01-701-1015 WEIGHT          9       NA     7
#>  9 01-701-1015 WEIGHT         10       NA     8
#> 10 01-701-1015 WEIGHT         11       NA     9
#> # … with 2,294 more rows