Add a First Event Parameter

Add a new parameter for the first event occurring in a dataset. AVALC and AVAL indicate if an event occurred and ADT is set to the date of the first event. For example, the function can derive a parameter for the first disease progression.

Usage

derive_param_first_event(
  dataset,
  dataset_adsl,
  dataset_source,
  filter_source,
  date_var,
  subject_keys = vars(STUDYID, USUBJID),
  set_values_to,
  check_type = "warning"
)

Arguments

dataset

Input dataset

The PARAMCD variable is expected.

dataset_adsl

ADSL input dataset

The variables specified for subject_keys are expected. For each observation of the specified dataset a new observation is added to the input dataset.

dataset_source

Source dataset

All observations in the specified dataset fulfilling the condition specified by filter_source are considered as event.

The variables specified by the subject_keys and date_var parameter are expected.

filter_source

Source filter

All observations in dataset_source fulfilling the specified condition are considered as event.

For subjects with at least one event AVALC is set to "Y", AVAL to 1, and ADT to the first date where the condition is fulfilled.

For all other subjects AVALC is set to "N", AVAL to 0, and ADT to NA.

date_var

Date variable

Date variable in the source dataset (dataset_source). The variable is used to sort the source dataset. ADT is set to the specified variable for events.

subject_keys

Variables to uniquely identify a subject

A list of symbols created using vars() is expected.

set_values_to

Variables to set

A named list returned by vars() defining the variables to be set for the new parameter, e.g. vars(PARAMCD = "PD", PARAM = "Disease Progression") is expected. The values must be symbols, character strings, numeric values, or NA.

check_type

Check uniqueness?

If "warning" or "error" is specified, a message is issued if the observations of the input dataset restricted to the source parameter (source_param) are not unique with respect to the subject keys (subject_key parameter) and ADT.

Default: "warning"

Permitted Values: "none", "warning", "error"

Value

The input dataset with a new parameter indicating if and when an event occurred

Details

The input dataset is restricted to observations fulfilling filter_source.

For each subject (with respect to the variables specified for the subject_keys parameter) the first observation (with respect to date_var) where the event condition (filter_source parameter) is fulfilled is selected.

For each observation in dataset_adsl a new observation is created. For subjects with event AVALC is set to "Y", AVAL to 1, and ADT to the first date where the event condition is fulfilled. For all other subjects AVALC is set to "N", AVAL to 0, and ADT to NA. For subjects with event all variables from dataset_source are kept. For subjects without event all variables which are in both dataset_adsl and dataset_source are kept.

The variables specified by the set_values_to parameter are added to the new observations.

The new observations are added to input dataset.

Author

Stefan Bundfuss

Examples

library(dplyr)
library(lubridate)

# Derive a new parameter for the first disease progression (PD)
adsl <- tibble::tribble(
  ~USUBJID, ~DTHDT,
  "1",      ymd("2022-05-13"),
  "2",      ymd(""),
  "3",      ymd("")
) %>%
  mutate(STUDYID = "XX1234")

adrs <- tibble::tribble(
  ~USUBJID, ~ADTC,        ~AVALC,
  "1",      "2020-01-02", "PR",
  "1",      "2020-02-01", "CR",
  "1",      "2020-03-01", "CR",
  "1",      "2020-04-01", "SD",
  "2",      "2021-06-15", "SD",
  "2",      "2021-07-16", "PD",
  "2",      "2021-09-14", "PD"
) %>%
  mutate(
    STUDYID = "XX1234",
    ADT = ymd(ADTC),
    PARAMCD = "OVR",
    PARAM = "Overall Response",
    ANL01FL = "Y"
  ) %>%
  select(-ADTC)

derive_param_first_event(
  adrs,
  dataset_adsl = adsl,
  dataset_source = adrs,
  filter_source = PARAMCD == "OVR" & AVALC == "PD",
  date_var = ADT,
  set_values_to = vars(
    PARAMCD = "PD",
    PARAM = "Disease Progression",
    ANL01FL = "Y"
  )
)
#> Joining, by = c("USUBJID", "STUDYID")
#> # A tibble: 10 x 8
#>    USUBJID AVALC STUDYID ADT        PARAMCD PARAM               ANL01FL  AVAL
#>    <chr>   <chr> <chr>   <date>     <chr>   <chr>               <chr>   <dbl>
#>  1 1       PR    XX1234  2020-01-02 OVR     Overall Response    Y          NA
#>  2 1       CR    XX1234  2020-02-01 OVR     Overall Response    Y          NA
#>  3 1       CR    XX1234  2020-03-01 OVR     Overall Response    Y          NA
#>  4 1       SD    XX1234  2020-04-01 OVR     Overall Response    Y          NA
#>  5 2       SD    XX1234  2021-06-15 OVR     Overall Response    Y          NA
#>  6 2       PD    XX1234  2021-07-16 OVR     Overall Response    Y          NA
#>  7 2       PD    XX1234  2021-09-14 OVR     Overall Response    Y          NA
#>  8 2       Y     XX1234  2021-07-16 PD      Disease Progression Y           1
#>  9 1       N     XX1234  NA         PD      Disease Progression Y           0
#> 10 3       N     XX1234  NA         PD      Disease Progression Y           0

# derive parameter indicating death
derive_param_first_event(
  dataset = adrs,
  dataset_adsl = adsl,
  dataset_source = adsl,
  filter_source = !is.na(DTHDT),
  date_var = DTHDT,
  set_values_to = vars(
    PARAMCD = "DEATH",
    PARAM = "Death",
    ANL01FL = "Y"
  )
)
#> Joining, by = c("USUBJID", "STUDYID")
#> # A tibble: 10 x 9
#>    USUBJID AVALC STUDYID ADT        PARAMCD PARAM       ANL01FL DTHDT       AVAL
#>    <chr>   <chr> <chr>   <date>     <chr>   <chr>       <chr>   <date>     <dbl>
#>  1 1       PR    XX1234  2020-01-02 OVR     Overall Re… Y       NA            NA
#>  2 1       CR    XX1234  2020-02-01 OVR     Overall Re… Y       NA            NA
#>  3 1       CR    XX1234  2020-03-01 OVR     Overall Re… Y       NA            NA
#>  4 1       SD    XX1234  2020-04-01 OVR     Overall Re… Y       NA            NA
#>  5 2       SD    XX1234  2021-06-15 OVR     Overall Re… Y       NA            NA
#>  6 2       PD    XX1234  2021-07-16 OVR     Overall Re… Y       NA            NA
#>  7 2       PD    XX1234  2021-09-14 OVR     Overall Re… Y       NA            NA
#>  8 1       Y     XX1234  2022-05-13 DEATH   Death       Y       2022-05-13     1
#>  9 2       N     XX1234  NA         DEATH   Death       Y       NA             0
#> 10 3       N     XX1234  NA         DEATH   Death       Y       NA             0