Derive Last Dose with User-Defined Groupings

Add a variable for user-defined dose grouping of the last dose to the input dataset.

derive_var_last_dose_grp(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  by_vars = vars(STUDYID, USUBJID),
  dose_id = vars(),
  dose_date,
  analysis_date,
  single_dose_condition = (EXDOSFRQ == "ONCE"),
  new_var,
  grp_brks,
  grp_lbls,
  include_lowest = TRUE,
  right = TRUE,
  dose_var = EXDOSE,
  traceability_vars = NULL
)

Arguments

dataset
Input dataset. The variables specified by the by_vars and analysis_date parameters are expected.

dataset_ex
Input EX dataset. The variables specified by the by_vars, dose_date, new_vars parameters, and source variables from traceability_vars parameter are expected.

filter_ex
Filtering condition applied to EX dataset. For example, it can be used to filter for valid dose. Defaults to NULL.

by_vars
Variables to join by (created by dplyr::vars).

dose_id
Variables to identify unique dose (created by dplyr::vars). Defaults to empty vars().

dose_date
The EX dose date variable.

analysis_date
The analysis date variable.

single_dose_condition
The condition for checking if dataset_ex is single dose. An error is issued if the condition is not true. Defaults to (EXDOSFRQ == "ONCE").

new_var
The output variable defined by the user.

grp_brks
User supplied breaks to apply to groups. Refer to breaks parameter in cut() for details.

grp_lbls
User supplied labels to apply to groups. Refer to labels parameter in cut() for details.

include_lowest
logical, indicating if a value equal to the lowest (or highest, for right = FALSE) ‘breaks’ value should be included. Refer to include.lowest parameter in cut() for details.

right
Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa. Refer to right parameter in cut() for details.

dose_var
The source dose amount variable. Defaults to EXDOSE.

traceability_vars
A named list returned by vars() listing the traceability variables, e.g. vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ). The left-hand side (names of the list elements) gives the names of the traceability variables in the returned dataset. The right-hand side (values of the list elements) gives the values of the traceability variables in the returned dataset. These can be either strings or symbols referring to existing variables.

Value

Input dataset with additional column new_var.

Details

Last dose is the dose with maximum dose_date that is lower to or equal to the analysis_date per by_vars for each observation in dataset. The last dose group is then derived by user-defined grouping, which groups dose_var as specified in grp_brks, and returns grp_lbls as the values for new_var.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data(ae)
data(ex_single)

ae %>%
   head(100) %>%
   derive_var_last_dose_grp(
   head(ex_single, 100),
   filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
   nchar(EXENDTC) >= 10,
   by_vars = vars(STUDYID, USUBJID),
   dose_date = EXSTDTC,
   new_var = LDGRP,
   grp_brks = c(0, 20, 40, 60),
   grp_lbls = c("Low", "Medium", "High"),
   include_lowest = TRUE,
   right = TRUE,
   dose_var = EXDOSE,
   analysis_date = AESTDTC,
   single_dose_condition = (EXSTDTC == EXENDTC),
   traceability_vars = dplyr::vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ, LDOSEVAR = "EXENDTC")
   ) %>%
   select(USUBJID, LDGRP, LDOSEDOM, LDOSESEQ, LDOSEVAR)

#> # A tibble: 100 x 5
#>    USUBJID     LDGRP LDOSEDOM LDOSESEQ LDOSEVAR
#>    <chr>       <chr> <chr>       <int> <chr>   
#>  1 01-701-1015 <NA>  <NA>           NA <NA>    
#>  2 01-701-1015 <NA>  <NA>           NA <NA>    
#>  3 01-701-1015 <NA>  <NA>           NA <NA>    
#>  4 01-701-1023 Low   EX              7 EXENDTC 
#>  5 01-701-1023 Low   EX              3 EXENDTC 
#>  6 01-701-1023 Low   EX              3 EXENDTC 
#>  7 01-701-1023 Low   EX              3 EXENDTC 
#>  8 01-701-1028 <NA>  <NA>           NA <NA>    
#>  9 01-701-1028 <NA>  <NA>           NA <NA>    
#> 10 01-701-1034 <NA>  <NA>           NA <NA>    
#> # … with 90 more rows