Derive Last Known Alive Date

Add the last known alive date (LSTALVDT) to the dataset.

derive_var_lstalvdt(dataset, ..., subject_keys = vars(STUDYID, USUBJID))

Arguments

dataset
Input dataset

The USUBJID variable is expected.

...
Source of known alive dates. A lstalvdt_source() object is expected.

subject_keys
Variables to uniquely identify a subject

A list of quosures where the expressions are symbols as returned by vars() is expected.

Value

The input dataset with the LSTALVDT variable added.

Details

The following steps are performed to create the output dataset:

For each source dataset the observations as specified by the filter element are selected. Then for each patient the last observation (with respect to date) is selected.

The LSTALVDT variable is set to the variable specified by the date element. If the date variable is a datetime variable, only the datepart is copied. If the source variable is a character variable, it is converted to a date. If the date is incomplete, it is imputed as specified by the date_imputation element.

The variables specified by the traceability_vars element are added.

The selected observations of all source datasets are combined into a single dataset.

For each patient the last observation (with respect to the LSTALVDT variable) from the single dataset is selected and the new variable is merged to the input dataset.

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data("dm")
data("ae")
data("lb")
data("adsl")

ae_start <- lstalvdt_source(
  dataset = ae,
  date = AESTDTC,
  date_imputation = "first"
)
ae_end <- lstalvdt_source(
  dataset = ae,
  date = AEENDTC,
  date_imputation = "first"
)
lb_date <- lstalvdt_source(
  dataset = lb,
  date = LBDTC,
  filter = nchar(LBDTC) >= 10
)
adsl_date <- lstalvdt_source(dataset = adsl, date = TRTEDT)

dm %>%
  derive_var_lstalvdt(ae_start, ae_end, lb_date, adsl_date) %>%
  select(USUBJID, LSTALVDT)

#> # A tibble: 306 x 2
#>    USUBJID     LSTALVDT  
#>    <chr>       <date>    
#>  1 01-701-1015 2014-07-02
#>  2 01-701-1023 2012-09-02
#>  3 01-701-1028 2014-01-14
#>  4 01-701-1033 2014-04-14
#>  5 01-701-1034 2014-12-30
#>  6 01-701-1047 2013-04-07
#>  7 01-701-1057 NA        
#>  8 01-701-1097 2014-07-09
#>  9 01-701-1111 2012-09-17
#> 10 01-701-1115 2013-01-23
#> # … with 296 more rows

# derive last alive date and traceability variables
ae_start <- lstalvdt_source(
  dataset = ae,
  date = AESTDTC,
  date_imputation = "first",
  traceability_vars = vars(
    LALVDOM = "AE",
    LALVSEQ = AESEQ,
    LALVVAR = "AESTDTC"
  )
)

ae_end <- lstalvdt_source(
  dataset = ae,
  date = AEENDTC,
  date_imputation = "first",
  traceability_vars = vars(
    LALVDOM = "AE",
    LALVSEQ = AESEQ,
    LALVVAR = "AEENDTC"
  )
)
lb_date <- lstalvdt_source(
  dataset = lb,
  date = LBDTC,
  filter = nchar(LBDTC) >= 10,
  traceability_vars = vars(
    LALVDOM = "LB",
    LALVSEQ = LBSEQ,
    LALVVAR = "LBDTC"
  )
)

adsl_date <- lstalvdt_source(
  dataset = adsl,
  date = TRTEDT,
  traceability_vars = vars(
    LALVDOM = "ADSL",
    LALVSEQ = NA_integer_,
    LALVVAR = "TRTEDTM"
  )
)

dm %>%
  derive_var_lstalvdt(ae_start, ae_end, lb_date, adsl_date) %>%
  select(USUBJID, LSTALVDT, LALVDOM, LALVSEQ, LALVVAR)

#> # A tibble: 306 x 5
#>    USUBJID     LSTALVDT   LALVDOM LALVSEQ LALVVAR
#>    <chr>       <date>     <chr>     <dbl> <chr>  
#>  1 01-701-1015 2014-07-02 ADSL         NA TRTEDTM
#>  2 01-701-1023 2012-09-02 LB          107 LBDTC  
#>  3 01-701-1028 2014-01-14 ADSL         NA TRTEDTM
#>  4 01-701-1033 2014-04-14 LB          107 LBDTC  
#>  5 01-701-1034 2014-12-30 ADSL         NA TRTEDTM
#>  6 01-701-1047 2013-04-07 LB          134 LBDTC  
#>  7 01-701-1057 NA         <NA>         NA <NA>   
#>  8 01-701-1097 2014-07-09 ADSL         NA TRTEDTM
#>  9 01-701-1111 2012-09-17 LB           73 LBDTC  
#> 10 01-701-1115 2013-01-23 ADSL         NA TRTEDTM
#> # … with 296 more rows