Derive Total Treatment Duration (days)

Derives total treatment duration (days) (TRTDURD)

derive_var_trtdurd(dataset, start_date = TRTSDT, end_date = TRTEDT)

Arguments

dataset
Input dataset

The columns specified by the start_date and the end_date parameter are expected.

start_date
The start date

A date or date-time object is expected.

Default: TRTSDT

end_date
The end date

A date or date-time object is expected.

Default: TRTEDT

Value

The input dataset with TRTDURD added

Details

The total treatment duration is derived as the number of days from start to end date plus one.

See also

Examples

data <- tibble::tribble(
  ~TRTSDT, ~TRTEDT,
  lubridate::ymd('2020-01-01'), lubridate::ymd('2020-02-24'))

derive_var_trtdurd(data)

#> # A tibble: 1 x 3
#>   TRTSDT     TRTEDT     TRTDURD
#>   <date>     <date>       <dbl>
#> 1 2020-01-01 2020-02-24      55