Merge a (Imputed) Datetime Variable

Merge a imputed datetime variable, date imputation flag, and time imputation flag from a dataset to the input dataset. The observations to merge can be selected by a condition and/or selecting the first or last observation for each by group.

Usage

derive_vars_merged_dtm(
  dataset,
  dataset_add,
  by_vars,
  order = NULL,
  new_vars_prefix,
  filter_add = NULL,
  mode = NULL,
  dtc,
  date_imputation = NULL,
  time_imputation = "00:00:00",
  flag_imputation = "auto",
  min_dates = NULL,
  max_dates = NULL,
  preserve = FALSE,
  check_type = "warning",
  duplicate_msg = NULL
)

Arguments

dataset

Input dataset

The variables specified by the by_vars parameter are expected.

dataset_add

Additional dataset

The variables specified by the by_vars, the dtc, and the order parameter are expected.

by_vars

Grouping variables

The input dataset and the selected observations from the additional dataset are merged by the specified by variables. The by variables must be a unique key of the selected observations.

Permitted Values: list of variables created by vars()

order

Sort order

If the parameter is set to a non-null value, for each by group the first or last observation from the additional dataset is selected with respect to the specified order. The imputed datetime variable can be specified as well (see examples below).

Default: NULL

Permitted Values: list of variables or desc(<variable>) function calls created by vars(), e.g., vars(ADT, desc(AVAL) or NULL

new_vars_prefix

Prefix used for the output variable(s).

A character scalar is expected. For the date variable "DT" is appended to the specified prefix, for the date imputation flag "DTF", and for the time imputation flag "TMF". I.e., for new_vars_prefix = "AST" the variables ASTDT, ASTDTF, and ASTTMF are created.

filter_add

Filter for additional dataset (dataset_add)

Only observations fulfilling the specified condition are taken into account for merging. If the parameter is not specified, all observations are considered.

Default: NULL

Permitted Values: a condition

mode

Selection mode

Determines if the first or last observation is selected. If the order parameter is specified, mode must be non-null.

If the order parameter is not specified, the mode parameter is ignored.

Default: NULL

Permitted Values: "first", "last", NULL

dtc

The '--DTC' date to impute

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss. If the year part is not recorded (missing date), no imputation is performed.

date_imputation

The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a

format with month and day specified as "mm-dd": e.g. "06-15" for the 15th of June,

or as a keyword: "FIRST", "MID", "LAST" to impute to the first/mid/last day/month.

Default is NULL.

time_imputation

The value to impute the time when a timepart is missing.

A character value is expected, either as a

format with hour, min and sec specified as "hh:mm:ss": e.g. "00:00:00" for the start of the day,

or as a keyword: "FIRST","LAST" to impute to the start/end of a day.

Default is "00:00:00".

flag_imputation

Whether the date/time imputation flag(s) must also be derived.

If "auto" is specified, the date imputation flag is derived if the date_imputation parameter is not null and the time imputation flag is derived if the time_imputation parameter is not null

Default: "auto"

Permitted Values: "auto", "date", "time", "both", or "none"

min_dates

Minimum dates

A list of dates is expected. It is ensured that the imputed date is not before any of the specified dates, e.g., that the imputed adverse event start date is not before the first treatment date. Only dates which are in the range of possible dates of the dtc value are considered. The possible dates are defined by the missing parts of the dtc date (see example below). This ensures that the non-missing parts of the dtc date are not changed. A date or date-time object is expected. For example

impute_dtc(
"2020-11",
min_dates = list(
 ymd_hms("2020-12-06T12:12:12"),
 ymd_hms("2020-11-11T11:11:11")
),
date_imputation = "first"
)

returns "2020-11-11T11:11:11" because the possible dates for "2020-11" range from "2020-11-01T00:00:00" to "2020-11-30T23:59:59". Therefore "2020-12-06T12:12:12" is ignored. Returning "2020-12-06T12:12:12" would have changed the month although it is not missing (in the dtc date).

max_dates

Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after any of the specified dates, e.g., that the imputed date is not after the data cut off date. Only dates which are in the range of possible dates are considered. A date or date-time object is expected.

preserve

Preserve day if month is missing and day is present

For example "2019---07" would return "2019-06-07 if preserve = TRUE (and date_imputation = "MID").

Permitted Values: TRUE, FALSE

Default: FALSE

check_type

Check uniqueness?

If "warning" or "error" is specified, the specified message is issued if the observations of the (restricted) additional dataset are not unique with respect to the by variables and the order.

Default: "warning"

Permitted Values: "none", "warning", "error"

duplicate_msg

Message of unique check

If the uniqueness check fails, the specified message is displayed.

Default:

paste("Dataset `dataset_add` contains duplicate records with respect to",
      enumerate(vars2chr(by_vars)))

Value

The output dataset contains all observations and variables of the input dataset and additionally the variable <new_vars_prefix>DT and optionally the variables <new_vars_prefix>DTF and <new_vars_prefix>TMFderived from the additional dataset (dataset_add).

Details

The additional dataset is restricted to the observations matching the filter_add condition.

The datetime variable and if requested, the date imputation flag and time imputation flag is added to the additional dataset.

If order is specified, for each by group the first or last observation (depending on mode) is selected.

The date and flag variables are merged to the input dataset.

Author

Stefan Bundfuss

Examples

library(admiral.test)
library(dplyr, warn.conflicts = FALSE)
data("admiral_dm")
data("admiral_ex")

# derive treatment start datetime (TRTSDTM)
derive_vars_merged_dtm(
  select(admiral_dm, STUDYID, USUBJID),
  dataset_add = admiral_ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTS",
  dtc = EXSTDTC,
  date_imputation = "first",
  time_imputation = "first",
  order = vars(TRTSDTM),
  mode = "first"
)
#> # A tibble: 306 x 5
#>    STUDYID      USUBJID     TRTSDTM             TRTSDTF TRTSTMF
#>    <chr>        <chr>       <dttm>              <chr>   <chr>  
#>  1 CDISCPILOT01 01-701-1015 2014-01-02 00:00:00 NA      H      
#>  2 CDISCPILOT01 01-701-1023 2012-08-05 00:00:00 NA      H      
#>  3 CDISCPILOT01 01-701-1028 2013-07-19 00:00:00 NA      H      
#>  4 CDISCPILOT01 01-701-1033 2014-03-18 00:00:00 NA      H      
#>  5 CDISCPILOT01 01-701-1034 2014-07-01 00:00:00 NA      H      
#>  6 CDISCPILOT01 01-701-1047 2013-02-12 00:00:00 NA      H      
#>  7 CDISCPILOT01 01-701-1057 NA                  NA      NA     
#>  8 CDISCPILOT01 01-701-1097 2014-01-01 00:00:00 NA      H      
#>  9 CDISCPILOT01 01-701-1111 2012-09-07 00:00:00 NA      H      
#> 10 CDISCPILOT01 01-701-1115 2012-11-30 00:00:00 NA      H      
#> # … with 296 more rows

# derive treatment end datetime (TRTEDTM) (without date imputation)
derive_vars_merged_dtm(
  select(admiral_dm, STUDYID, USUBJID),
  dataset_add = admiral_ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTE",
  dtc = EXENDTC,
  time_imputation = "last",
  order = vars(TRTEDTM),
  mode = "last"
)
#> # A tibble: 306 x 4
#>    STUDYID      USUBJID     TRTEDTM             TRTETMF
#>    <chr>        <chr>       <dttm>              <chr>  
#>  1 CDISCPILOT01 01-701-1015 2014-07-02 23:59:59 H      
#>  2 CDISCPILOT01 01-701-1023 2012-09-01 23:59:59 H      
#>  3 CDISCPILOT01 01-701-1028 2014-01-14 23:59:59 H      
#>  4 CDISCPILOT01 01-701-1033 2014-03-31 23:59:59 H      
#>  5 CDISCPILOT01 01-701-1034 2014-12-30 23:59:59 H      
#>  6 CDISCPILOT01 01-701-1047 2013-03-09 23:59:59 H      
#>  7 CDISCPILOT01 01-701-1057 NA                  NA     
#>  8 CDISCPILOT01 01-701-1097 2014-07-09 23:59:59 H      
#>  9 CDISCPILOT01 01-701-1111 2012-09-16 23:59:59 H      
#> 10 CDISCPILOT01 01-701-1115 2013-01-23 23:59:59 H      
#> # … with 296 more rows