Optional Filter

Filters the input dataset if the provided expression is not NULL

Usage

filter_if(dataset, filter)

Arguments

dataset

Input dataset

filter

A filter condition. Must be a quosure.

Value

A data.frame containing all rows in dataset matching filter or just dataset if filter is NULL

Author

Thomas Neitmann

Examples

library(admiral.test)
data(admiral_vs)

admiral::filter_if(admiral_vs, rlang::quo(NULL))
#> # A tibble: 29,643 x 24
#>    STUDYID DOMAIN USUBJID VSSEQ VSTESTCD VSTEST VSPOS VSORRES VSORRESU VSSTRESC
#>    <chr>   <chr>  <chr>   <dbl> <chr>    <chr>  <chr> <chr>   <chr>    <chr>   
#>  1 CDISCP… VS     01-701…     1 DIABP    Diast… SUPI… 64      mmHg     64      
#>  2 CDISCP… VS     01-701…     2 DIABP    Diast… STAN… 83      mmHg     83      
#>  3 CDISCP… VS     01-701…     3 DIABP    Diast… STAN… 57      mmHg     57      
#>  4 CDISCP… VS     01-701…     4 DIABP    Diast… SUPI… 68      mmHg     68      
#>  5 CDISCP… VS     01-701…     5 DIABP    Diast… STAN… 59      mmHg     59      
#>  6 CDISCP… VS     01-701…     6 DIABP    Diast… STAN… 71      mmHg     71      
#>  7 CDISCP… VS     01-701…     7 DIABP    Diast… SUPI… 56      mmHg     56      
#>  8 CDISCP… VS     01-701…     8 DIABP    Diast… STAN… 51      mmHg     51      
#>  9 CDISCP… VS     01-701…     9 DIABP    Diast… STAN… 61      mmHg     61      
#> 10 CDISCP… VS     01-701…    10 DIABP    Diast… SUPI… 67      mmHg     67      
#> # … with 29,633 more rows, and 14 more variables: VSSTRESN <dbl>,
#> #   VSSTRESU <chr>, VSSTAT <chr>, VSLOC <chr>, VSBLFL <chr>, VISITNUM <dbl>,
#> #   VISIT <chr>, VISITDY <dbl>, VSDTC <chr>, VSDY <dbl>, VSTPT <chr>,
#> #   VSTPTNUM <dbl>, VSELTM <chr>, VSTPTREF <chr>
admiral::filter_if(admiral_vs, rlang::quo(VSTESTCD == "WEIGHT"))
#> # A tibble: 2,050 x 24
#>    STUDYID DOMAIN USUBJID VSSEQ VSTESTCD VSTEST VSPOS VSORRES VSORRESU VSSTRESC
#>    <chr>   <chr>  <chr>   <dbl> <chr>    <chr>  <chr> <chr>   <chr>    <chr>   
#>  1 CDISCP… VS     01-701…   142 WEIGHT   Weight ""    119.0   LB       53.98   
#>  2 CDISCP… VS     01-701…   143 WEIGHT   Weight ""    120.0   LB       54.43   
#>  3 CDISCP… VS     01-701…   144 WEIGHT   Weight ""    117.0   LB       53.07   
#>  4 CDISCP… VS     01-701…   145 WEIGHT   Weight ""    119.0   LB       53.98   
#>  5 CDISCP… VS     01-701…   146 WEIGHT   Weight ""    117.0   LB       53.07   
#>  6 CDISCP… VS     01-701…   147 WEIGHT   Weight ""    117.0   LB       53.07   
#>  7 CDISCP… VS     01-701…   148 WEIGHT   Weight ""    117.0   LB       53.07   
#>  8 CDISCP… VS     01-701…   149 WEIGHT   Weight ""    117.0   LB       53.07   
#>  9 CDISCP… VS     01-701…   150 WEIGHT   Weight ""    117.0   LB       53.07   
#> 10 CDISCP… VS     01-701…   151 WEIGHT   Weight ""    117.0   LB       53.07   
#> # … with 2,040 more rows, and 14 more variables: VSSTRESN <dbl>,
#> #   VSSTRESU <chr>, VSSTAT <chr>, VSLOC <chr>, VSBLFL <chr>, VISITNUM <dbl>,
#> #   VISIT <chr>, VISITDY <dbl>, VSDTC <chr>, VSDY <dbl>, VSTPT <chr>,
#> #   VSTPTNUM <dbl>, VSELTM <chr>, VSTPTREF <chr>