Default Format for the Disposition Reason

Define a function to map the disposition reason, to be used as a parameter in derive_vars_disposition_reason().

Usage

format_reason_default(reason, reason_spe = NULL)

Arguments

reason

the disposition variable used for the mapping (e.g. DSDECOD).

reason_spe

the disposition variable used for the mapping of the details if required (e.g. DSTERM).

Value

A character vector

Details

format_reason_default(DSDECOD) returns DSDECOD when DSDECOD is not 'COMPLETED' nor NA.
format_reason_default(DSDECOD, DSTERM) returns DSTERM when DSDECOD is equal to 'OTHER'.
Usually this function can not be used with %>%.

See also

Author

Samia Kabi

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data("admiral_dm")
data("admiral_ds")

# Derive DCSREAS using format_reason_default
admiral_dm %>%
  derive_vars_disposition_reason(
    dataset_ds = admiral_ds,
    new_var = DCSREAS,
    reason_var = DSDECOD,
    format_new_vars = format_reason_default,
    filter_ds = DSCAT == "DISPOSITION EVENT"
  ) %>%
  select(STUDYID, USUBJID, DCSREAS)
#> # A tibble: 306 x 3
#>    STUDYID      USUBJID     DCSREAS                    
#>    <chr>        <chr>       <chr>                      
#>  1 CDISCPILOT01 01-701-1015 NA                         
#>  2 CDISCPILOT01 01-701-1023 ADVERSE EVENT              
#>  3 CDISCPILOT01 01-701-1028 NA                         
#>  4 CDISCPILOT01 01-701-1033 STUDY TERMINATED BY SPONSOR
#>  5 CDISCPILOT01 01-701-1034 NA                         
#>  6 CDISCPILOT01 01-701-1047 ADVERSE EVENT              
#>  7 CDISCPILOT01 01-701-1057 SCREEN FAILURE             
#>  8 CDISCPILOT01 01-701-1097 NA                         
#>  9 CDISCPILOT01 01-701-1111 ADVERSE EVENT              
#> 10 CDISCPILOT01 01-701-1115 ADVERSE EVENT              
#> # … with 296 more rows