Create an lstalvdt_source object

Deprecated, please use date_source() instead.

Usage

lstalvdt_source(
  dataset_name,
  filter = NULL,
  date,
  date_imputation = NULL,
  traceability_vars = NULL
)

Arguments

dataset_name

The name of the dataset, i.e. a string, used to search for the last known alive date.

filter

An unquoted condition for filtering dataset.

date

A variable providing a date where the patient was known to be alive. A date, a datetime, or a character variable containing ISO 8601 dates can be specified. An unquoted symbol is expected.

date_imputation

A string defining the date imputation for date. See date_imputation parameter of impute_dtc() for valid values.

traceability_vars

A named list returned by vars() defining the traceability variables, e.g. vars(LALVDOM = "AE", LALVSEQ = AESEQ, LALVVAR = "AESTDTC"). The values must be a symbol, a character string, or NA.

Value

An object of class lstalvdt_source.

Author

Stefan Bundfuss