Negate List of Variables

The function adds a minus sign as prefix to each variable.

Usage

negate_vars(vars = NULL)

Arguments

vars

List of variables created by exprs()

Default value

NULL

Value

A list of expressions

Details

This is useful if a list of variables should be removed from a dataset, e.g., select(!!!negate_vars(by_vars)) removes all by variables.

See also

Examples

negate_vars(exprs(USUBJID, STUDYID))
#> [[1]]
#> -USUBJID
#> 
#> [[2]]
#> -STUDYID
#>