Create a Set of Parameters

Create a set of variable parameters/function arguments to be used in call_derivation().

Usage

params(...)

Arguments

...

One or more named arguments

Value

An object of class params

See also

Author

Thomas Neitmann, Tracey Wang

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data(admiral_ae)
data(admiral_adsl)

adae <- admiral_ae[sample(1:nrow(admiral_ae), 1000), ] %>%
  select(USUBJID, AESTDTC, AEENDTC) %>%
  derive_vars_merged(
    dataset_add = admiral_adsl,
    new_vars = vars(TRTSDT, TRTEDT),
    by_vars = vars(USUBJID)
  )

## In order to derive both `ASTDT` and `AENDT` in `ADAE`, one can use `derive_vars_dt()`
adae %>%
  derive_vars_dt(
    new_vars_prefix = "AST",
    dtc = AESTDTC,
    date_imputation = "first",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  ) %>%
  derive_vars_dt(
    new_vars_prefix = "AEN",
    dtc = AEENDTC,
    date_imputation = "last",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  )
#> # A tibble: 1,000 x 7
#>    USUBJID     AESTDTC   AEENDTC     TRTSDT     TRTEDT     ASTDT      AENDT     
#>    <chr>       <chr>     <chr>       <date>     <date>     <date>     <date>    
#>  1 01-704-1120 2013-12-… ""          2013-12-02 2014-02-01 2013-12-23 NA        
#>  2 01-718-1254 2013-07-… "2013-07-2… 2013-07-10 2014-01-09 2013-07-25 2013-07-25
#>  3 01-713-1179 2013-11-… "2013-12-0… 2013-08-09 2014-02-05 2013-11-23 2013-12-02
#>  4 01-711-1012 2013-04-… "2013-05-0… 2013-04-03 2013-04-29 2013-04-16 2013-05-02
#>  5 01-709-1029 2013-05-… "2013-05-1… 2012-12-25 2013-06-26 2013-05-12 2013-05-19
#>  6 01-704-1351 2013-10-… "2014-04-0… 2013-10-12 2014-04-18 2013-10-27 2014-04-04
#>  7 01-704-1323 2013-08-… ""          2013-07-08 2013-08-05 2013-08-04 NA        
#>  8 01-701-1275 2014-02-… "2014-06-0… 2014-02-07 2014-05-31 2014-02-24 2014-06-03
#>  9 01-701-1115 2013-01-… ""          2012-11-30 2013-01-23 2013-01-19 NA        
#> 10 01-704-1017 2013-11-… "2013-11-2… 2013-10-06 2013-11-18 2013-11-05 2013-11-22
#> # … with 990 more rows

## While `derive_vars_dt()` can only add one variable at a time, using `call_derivation()`
## one can add multiple variables in one go.
## The function arguments which are different from a variable to another (e.g. `new_vars_prefix`,
## `dtc`, and `date_imputation`) are specified as a list of `params()` in the `variable_params`
## argument of `call_derivation()`. All other arguments which are common to all variables
## (e.g. `min_dates` and `max_dates`) are specified outside of `variable_params` (i.e. in `...`).
call_derivation(
  dataset = adae,
  derivation = derive_vars_dt,
  variable_params = list(
    params(dtc = AESTDTC, date_imputation = "first", new_vars_prefix = "AST"),
    params(dtc = AEENDTC, date_imputation = "last", new_vars_prefix = "AEN")
  ),
  min_dates = vars(TRTSDT),
  max_dates = vars(TRTEDT)
)
#> # A tibble: 1,000 x 7
#>    USUBJID     AESTDTC   AEENDTC     TRTSDT     TRTEDT     ASTDT      AENDT     
#>    <chr>       <chr>     <chr>       <date>     <date>     <date>     <date>    
#>  1 01-704-1120 2013-12-… ""          2013-12-02 2014-02-01 2013-12-23 NA        
#>  2 01-718-1254 2013-07-… "2013-07-2… 2013-07-10 2014-01-09 2013-07-25 2013-07-25
#>  3 01-713-1179 2013-11-… "2013-12-0… 2013-08-09 2014-02-05 2013-11-23 2013-12-02
#>  4 01-711-1012 2013-04-… "2013-05-0… 2013-04-03 2013-04-29 2013-04-16 2013-05-02
#>  5 01-709-1029 2013-05-… "2013-05-1… 2012-12-25 2013-06-26 2013-05-12 2013-05-19
#>  6 01-704-1351 2013-10-… "2014-04-0… 2013-10-12 2014-04-18 2013-10-27 2014-04-04
#>  7 01-704-1323 2013-08-… ""          2013-07-08 2013-08-05 2013-08-04 NA        
#>  8 01-701-1275 2014-02-… "2014-06-0… 2014-02-07 2014-05-31 2014-02-24 2014-06-03
#>  9 01-701-1115 2013-01-… ""          2012-11-30 2013-01-23 2013-01-19 NA        
#> 10 01-704-1017 2013-11-… "2013-11-2… 2013-10-06 2013-11-18 2013-11-05 2013-11-22
#> # … with 990 more rows

## The above call using `call_derivation()` is equivalent to the call using `derive_vars_dt()`
## to derive variables `ASTDT` and `AENDT` separately at the beginning.