admiral: ADaM in R Asset Library

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package).

See also

Author

Maintainer: Thomas Neitmann thomas.neitmann@roche.com

Authors:

Stefan Bundfuss

Ben Straub

Samia Kabi

Gordon Miller

Teckla Akinyi

Andrew Smith

Konstantina Koukourikou

Ross Farrugia

Eric Simms

Annie Yang

Robin Koeger

Sophie Shapcott

Ojesh Upadhyay

Jack McGavigan

Kamila Duniec

Gayatri G

Alana Harris

Mahdi About

Pooja Kumari

Claudia Carlucci

Daniil Stefonishin

Other contributors:

Michael Thorpe [contributor]

Pavan Kumar [contributor]

Hamza Rahal [contributor]

Alice Ehmann [contributor]

Tom Ratford [contributor]

Vignesh Thanikachalam [contributor]

Ondrej Slama [contributor]

Shimeng Huang [contributor]

James Kim [contributor]

Shan Lee [contributor]

Bill Denney [contributor]

Syed Mubasheer [contributor]

Wenyi Liu [contributor]

Dinakar Kulkarni [contributor]

Tamara Senior [contributor]

Jordanna Morrish [contributor]

Anthony Howard [contributor]

Barbara O'Reilly [contributor]

John Kirkpatrick [contributor]

James Black [contributor]

Leena Khatri [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]