Call a Single Derivation Multiple Times

Call a single derivation multiple times with some parameters being fixed across iterations and others varying.

call_derivation(dataset = NULL, derivation, variable_params, ...)

Arguments

dataset
The input dataset

derivation
The derivation function to call

variable_params
A list of arguments that are different across iterations. Each set of arguments must be created using params().

...
Any number of named arguments that are fixed across iterations. If a parameter is specified both inside variable_params and ... then the value in variable_params overwrites the one in ...

Value

The input dataset with additional records/variables added depending on which derivation has been used.

See also

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
data(ae)
data(adsl)

adae <- ae[sample(1:nrow(ae), 1000), ] %>%
  left_join(adsl, by = "USUBJID") %>%
  select(USUBJID, AESTDTC, AEENDTC, TRTSDT, TRTEDT)

#> Warning: Column `USUBJID` has different attributes on LHS and RHS of join

## While `derive_vars_dt()` can only add one variable at a time, using `call_derivation()`
## one can add multiple variables in one go
call_derivation(
  dataset = adae,
  derivation = derive_vars_dt,
  variable_params = list(
    params(dtc = AESTDTC, date_imputation = "first", new_vars_prefix = "AST"),
    params(dtc = AEENDTC, date_imputation = "last", new_vars_prefix = "AEN")
  ),
  min_dates = vars(TRTSDT),
  max_dates = vars(TRTEDT)
)

#> # A tibble: 1,000 x 9
#>    USUBJID AESTDTC AEENDTC TRTSDT     TRTEDT     ASTDT      ASTDTF AENDT     
#>    <chr>   <chr>   <chr>   <date>     <date>     <date>     <chr>  <date>    
#>  1 01-707… 2014-0… "2014-… 2013-10-28 2014-04-26 2014-02-08 <NA>   2014-02-22
#>  2 01-701… 2013-0… ""      2013-05-20 2013-06-26 2013-06-26 <NA>   NA        
#>  3 01-708… 2014-0… "2014-… 2014-02-08 2014-03-21 2014-03-22 <NA>   2014-03-22
#>  4 01-708… 2014-0… ""      2013-12-26 2014-07-02 2014-06-16 <NA>   NA        
#>  5 01-714… 2014-0… "2014-… 2013-12-04 2014-06-17 2014-01-22 <NA>   2014-01-25
#>  6 01-701… 2013-0… ""      2013-01-05 2013-01-26 2013-01-19 <NA>   NA        
#>  7 01-708… 2013-0… "2013-… 2013-02-06 2013-03-22 2013-03-01 <NA>   2013-03-07
#>  8 01-708… 2014-0… ""      2013-08-23 2014-03-01 2014-01-29 <NA>   NA        
#>  9 01-701… 2013-1… "2013-… 2013-09-23 2014-03-16 2013-10-13 <NA>   2013-11-06
#> 10 01-713… 2014-0… "2014-… 2014-01-19 2014-05-16 2014-03-19 <NA>   2014-05-23
#> # … with 990 more rows, and 1 more variable: AENDTF <chr>

## The above call using `call_derivation()` is equivalent to the following
adae %>%
  derive_vars_dt(
    new_vars_prefix = "AST",
    dtc = AESTDTC,
    date_imputation = "first",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  ) %>%
  derive_vars_dt(
    new_vars_prefix = "AEN",
    dtc = AEENDTC,
    date_imputation = "last",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  )

#> # A tibble: 1,000 x 9
#>    USUBJID AESTDTC AEENDTC TRTSDT     TRTEDT     ASTDT      ASTDTF AENDT     
#>    <chr>   <chr>   <chr>   <date>     <date>     <date>     <chr>  <date>    
#>  1 01-707… 2014-0… "2014-… 2013-10-28 2014-04-26 2014-02-08 <NA>   2014-02-22
#>  2 01-701… 2013-0… ""      2013-05-20 2013-06-26 2013-06-26 <NA>   NA        
#>  3 01-708… 2014-0… "2014-… 2014-02-08 2014-03-21 2014-03-22 <NA>   2014-03-22
#>  4 01-708… 2014-0… ""      2013-12-26 2014-07-02 2014-06-16 <NA>   NA        
#>  5 01-714… 2014-0… "2014-… 2013-12-04 2014-06-17 2014-01-22 <NA>   2014-01-25
#>  6 01-701… 2013-0… ""      2013-01-05 2013-01-26 2013-01-19 <NA>   NA        
#>  7 01-708… 2013-0… "2013-… 2013-02-06 2013-03-22 2013-03-01 <NA>   2013-03-07
#>  8 01-708… 2014-0… ""      2013-08-23 2014-03-01 2014-01-29 <NA>   NA        
#>  9 01-701… 2013-1… "2013-… 2013-09-23 2014-03-16 2013-10-13 <NA>   2013-11-06
#> 10 01-713… 2014-0… "2014-… 2014-01-19 2014-05-16 2014-03-19 <NA>   2014-05-23
#> # … with 990 more rows, and 1 more variable: AENDTF <chr>