Call a Single Derivation Multiple Times

Call a single derivation multiple times with some parameters/arguments being fixed across iterations and others varying.

Usage

call_derivation(dataset = NULL, derivation, variable_params, ...)

Arguments

dataset

The input dataset

derivation

The derivation function to call

variable_params

A list of function arguments that are different across iterations. Each set of function arguments must be created using params().

...

Any number of named function arguments that stay the same across iterations. If a function argument is specified both inside variable_params and ... then the value in variable_params overwrites the one in ...

Value

The input dataset with additional records/variables added depending on which derivation has been used.

See also

Author

Thomas Neitmann, Stefan Bundfuss, Tracey Wang

Examples

library(dplyr, warn.conflicts = FALSE)
library(admiraltest)
data(admiral_ae)
data(admiral_adsl)

adae <-
  select(admiral_ae[sample(1:nrow(admiral_ae), 1000), ], USUBJID, AESTDTC, AEENDTC) %>%
  derive_vars_merged(
    dataset_add = admiral_adsl,
    new_vars = vars(TRTSDT, TRTEDT),
    by_vars = vars(USUBJID)
  )

## While `derive_vars_dt()` can only add one variable at a time, using `call_derivation()`
## one can add multiple variables in one go
call_derivation(
  dataset = adae,
  derivation = derive_vars_dt,
  variable_params = list(
    params(dtc = AESTDTC, date_imputation = "first", new_vars_prefix = "AST"),
    params(dtc = AEENDTC, date_imputation = "last", new_vars_prefix = "AEN")
  ),
  min_dates = vars(TRTSDT),
  max_dates = vars(TRTEDT)
)
#> # A tibble: 1,000 x 9
#>    USUBJID AESTDTC AEENDTC TRTSDT     TRTEDT     ASTDT      ASTDTF AENDT     
#>    <chr>   <chr>   <chr>   <date>     <date>     <date>     <chr>  <date>    
#>  1 01-711… 2007-10 ""      2013-04-03 2013-05-30 2007-10-01 D      NA        
#>  2 01-701… 2014-1… ""      2014-05-22 2014-11-20 2014-11-03 NA     NA        
#>  3 01-701… 2013-1… ""      2013-08-29 2013-11-05 2013-10-23 NA     NA        
#>  4 01-701… 2014-0… ""      2014-01-11 2014-07-10 2014-01-12 NA     NA        
#>  5 01-717… 2014-0… "2014-… 2013-09-01 2014-03-03 2014-01-13 NA     2014-01-14
#>  6 01-709… 2013-0… ""      2013-04-05 2013-10-02 2013-05-01 NA     NA        
#>  7 01-701… 2013-0… ""      2013-08-23 2014-02-20 2013-08-25 NA     NA        
#>  8 01-708… 2013-0… "2013-… 2013-04-20 2013-06-18 2013-06-13 NA     2013-06-25
#>  9 01-714… 2013-0… "2013-… 2013-04-26 2013-10-23 2013-05-12 NA     2013-07-17
#> 10 01-710… 2013-0… "2013-… 2013-02-10 2013-08-09 2013-04-16 NA     2013-04-16
#> # … with 990 more rows, and 1 more variable: AENDTF <chr>

## The above call using `call_derivation()` is equivalent to the following
adae %>%
  derive_vars_dt(
    new_vars_prefix = "AST",
    dtc = AESTDTC,
    date_imputation = "first",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  ) %>%
  derive_vars_dt(
    new_vars_prefix = "AEN",
    dtc = AEENDTC,
    date_imputation = "last",
    min_dates = vars(TRTSDT),
    max_dates = vars(TRTEDT)
  )
#> # A tibble: 1,000 x 9
#>    USUBJID AESTDTC AEENDTC TRTSDT     TRTEDT     ASTDT      ASTDTF AENDT     
#>    <chr>   <chr>   <chr>   <date>     <date>     <date>     <chr>  <date>    
#>  1 01-711… 2007-10 ""      2013-04-03 2013-05-30 2007-10-01 D      NA        
#>  2 01-701… 2014-1… ""      2014-05-22 2014-11-20 2014-11-03 NA     NA        
#>  3 01-701… 2013-1… ""      2013-08-29 2013-11-05 2013-10-23 NA     NA        
#>  4 01-701… 2014-0… ""      2014-01-11 2014-07-10 2014-01-12 NA     NA        
#>  5 01-717… 2014-0… "2014-… 2013-09-01 2014-03-03 2014-01-13 NA     2014-01-14
#>  6 01-709… 2013-0… ""      2013-04-05 2013-10-02 2013-05-01 NA     NA        
#>  7 01-701… 2013-0… ""      2013-08-23 2014-02-20 2013-08-25 NA     NA        
#>  8 01-708… 2013-0… "2013-… 2013-04-20 2013-06-18 2013-06-13 NA     2013-06-25
#>  9 01-714… 2013-0… "2013-… 2013-04-26 2013-10-23 2013-05-12 NA     2013-07-17
#> 10 01-710… 2013-0… "2013-… 2013-02-10 2013-08-09 2013-04-16 NA     2013-04-16
#> # … with 990 more rows, and 1 more variable: AENDTF <chr>