Derive the Date Imputation Flag

Derive the date imputation flag (*DTF) comparing a date character vector (--DTC) with a Date vector (*DT).

Usage

compute_dtf(dtc, dt)

Arguments

dtc

The date character vector (--DTC).

A character date is expected in a format like yyyy-mm-ddThh:mm:ss (partial or complete).

Default value

none

dt

The Date vector to compare.

A date object is expected.

Default value

none

Value

The date imputation flag (*DTF) (character value of "D", "M" , "Y" or NA)

Details

Usually this computation function can not be used with %>%.

See also

Examples

compute_dtf(dtc = "2019-07", dt = as.Date("2019-07-18"))
#> [1] "D"
compute_dtf(dtc = "2019", dt = as.Date("2019-07-18"))
#> [1] "M"
compute_dtf(dtc = "--06-01T00:00", dt = as.Date("2022-06-01"))
#> [1] "Y"
compute_dtf(dtc = "2022-06--T00:00", dt = as.Date("2022-06-01"))
#> [1] "D"
compute_dtf(dtc = "2022---01T00:00", dt = as.Date("2022-06-01"))
#> [1] "M"
compute_dtf(dtc = "2022----T00:00", dt = as.Date("2022-06-01"))
#> [1] "M"