Compute Corrected QT

Computes corrected QT using Bazett's, Fridericia's or Sagie's formula.

Usage

compute_qtc(qt, rr, method)

Arguments

qt

QT interval

A numeric vector is expected. It is expected that QT is measured in ms or msec.

rr

RR interval

A numeric vector is expected. It is expected that RR is measured in ms or msec.

method

Method used to QT correction

Permitted Values: "Bazett", "Fridericia", "Sagie"

Value

QT interval in ms

Details

Depending on the chosen method one of the following formulae is used.

Bazett: $$\frac{QT}{\sqrt{\frac{RR}{1000}}}$$

Fridericia: $$\frac{QT}{\sqrt[3]{\frac{RR}{1000}}}$$

Sagie: $$1000\left(\frac{QT}{1000} + 0.154\left(1 - \frac{RR}{1000}\right)\right)$$

Usually this computation function can not be used with %>%.

See also

Examples

compute_qtc(qt = 350, rr = 857, method = "Bazett")
#> [1] 378.0747

compute_qtc(qt = 350, rr = 857, method = "Fridericia")
#> [1] 368.4748

compute_qtc(qt = 350, rr = 857, method = "Sagie")
#> [1] 372.022