Derive the Time Imputation Flag

Derive the time imputation flag (*TMF) comparing a date character vector (--DTC) with a Datetime vector (*DTM).

Usage

compute_tmf(dtc, dtm, ignore_seconds_flag = FALSE)

Arguments

dtc

The date character vector (--DTC).

A character date is expected in a format like yyyy-mm-ddThh:mm:ss (partial or complete).

Default value

none

dtm

The Date vector to compare (*DTM).

A datetime object is expected.

Default value

none

ignore_seconds_flag

ADaM IG states that given SDTM (--DTC) variable, if only hours and minutes are ever collected, and seconds are imputed in (*DTM) as 00, then it is not necessary to set (*TMF) to "S". A user can set this to TRUE so the "S" Flag is dropped from (*TMF).

Please note that the default value of ignore_seconds_flag will change to TRUE in admiral 1.4.0.

Permitted values

"TRUE", "FALSE"

Default value

FALSE

Value

The time imputation flag (*TMF) (character value of "H", "M" , "S" or NA)

Details

Usually this computation function can not be used with %>%.

See also

Examples

library(lubridate)

compute_tmf(dtc = "2019-07-18T15:25", dtm = ymd_hms("2019-07-18T15:25:00"))
#> [1] "S"
compute_tmf(dtc = "2019-07-18T15", dtm = ymd_hms("2019-07-18T15:25:00"))
#> [1] "M"
compute_tmf(dtc = "2019-07-18", dtm = ymd("2019-07-18"))
#> [1] "H"
compute_tmf(dtc = "2022-05--T00:00", dtm = ymd_hms("2022-05-15T23:59:59"))
#> [1] "H"
compute_tmf(dtc = "2022-05--T23:00", dtm = ymd_hms("2022-05-15T23:59:59"))
#> [1] "M"
compute_tmf(dtc = "2022-05--T23:59:00", dtm = ymd_hms("2022-05-15T23:59:59"))
#> [1] "S"