Convert NAs Into Blank Strings

Turn NAs to blank strings .

Usage

convert_na_to_blanks(x)

# S3 method for default
convert_na_to_blanks(x)

# S3 method for character
convert_na_to_blanks(x)

# S3 method for list
convert_na_to_blanks(x)

# S3 method for data.frame
convert_na_to_blanks(x)

Arguments

x

Any R object

Value

An object of the same class as the input

Details

The default methods simply returns its input unchanged. The character method turns every instance of NA_character_ or NA into "" while preserving all attributes. When given a data frame as input the function keeps all non-character columns as is and applies the just described logic to character all attributes such as labels are preserved.

See also

Examples

library(tibble)

convert_na_to_blanks(c("a", "b", NA, "d", NA))
#> [1] "a" "b" ""  "d" "" 

df <- tribble(
  ~USUBJID,   ~RFICDTC,
  "1001", "2000-01-01",
  "1002", "2001-01-01",
  "1003",           NA
)
print(df)
#> # A tibble: 3 × 2
#>   USUBJID RFICDTC   
#>   <chr>   <chr>     
#> 1 1001    2000-01-01
#> 2 1002    2001-01-01
#> 3 1003    NA        
convert_na_to_blanks(df)
#> # A tibble: 3 × 2
#>   USUBJID RFICDTC     
#>   <chr>   <chr>       
#> 1 1001    "2000-01-01"
#> 2 1002    "2001-01-01"
#> 3 1003    ""