Derive a disposition date

Derive a disposition status date from the the relevant records in the disposition domain.

derive_disposition_dt(
  dataset,
  dataset_ds,
  new_var,
  dtc,
  filter_ds,
  date_imputation = NULL
)

Arguments

dataset
Input dataset

dataset_ds
Datasets containing the disposition information (e.g.: ds)

It must contains

STUDYID, USUBJID,

The variable(s) specified in the dtc

The variables used in filter_ds.

new_var
Name of the disposition date variable

a variable name is expected

dtc
The character date used to derive/impute the disposition date

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss. If the year part is not recorded (missing date), no imputation is performed.

filter_ds
Filter condition for the disposition data.

one observation per patient. An error is issued otherwise.

Permitted Values: logical expression.

date_imputation
The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a

format with day and month specified as 'dd-mm': e.g. '15-06' for the 15th of June

or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month

Default is NULL

Value

the input dataset with the disposition date (new_var) added

Examples

data("dm")
data("ds")
derive_disposition_dt(
  dataset = dm,
  dataset_ds = ds,
  new_var = FRVDT,
  dtc = DSSTDTC,
  filter_ds = DSCAT == "OTHER EVENT" & DSDECOD == "FINAL RETRIEVAL VISIT"
)

#> # A tibble: 306 x 26
#>    STUDYID DOMAIN USUBJID SUBJID RFSTDTC RFENDTC RFXSTDTC RFXENDTC RFICDTC
#>    <chr>   <chr>  <chr>   <chr>  <chr>   <chr>   <chr>    <chr>    <chr>  
#>  1 CDISCP… DM     01-701… 1015   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  2 CDISCP… DM     01-701… 1023   "2012-… "2012-… "2012-0… "2012-0… ""     
#>  3 CDISCP… DM     01-701… 1028   "2013-… "2014-… "2013-0… "2014-0… ""     
#>  4 CDISCP… DM     01-701… 1033   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  5 CDISCP… DM     01-701… 1034   "2014-… "2014-… "2014-0… "2014-1… ""     
#>  6 CDISCP… DM     01-701… 1047   "2013-… "2013-… "2013-0… "2013-0… ""     
#>  7 CDISCP… DM     01-701… 1057   ""      ""      ""       ""       ""     
#>  8 CDISCP… DM     01-701… 1097   "2014-… "2014-… "2014-0… "2014-0… ""     
#>  9 CDISCP… DM     01-701… 1111   "2012-… "2012-… "2012-0… "2012-0… ""     
#> 10 CDISCP… DM     01-701… 1115   "2012-… "2013-… "2012-1… "2013-0… ""     
#> # … with 296 more rows, and 17 more variables: RFPENDTC <chr>, DTHDTC <chr>,
#> #   DTHFL <chr>, SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>,
#> #   ETHNIC <chr>, ARMCD <chr>, ARM <chr>, ACTARMCD <chr>, ACTARM <chr>,
#> #   COUNTRY <chr>, DMDTC <chr>, DMDY <dbl>, FRVDT <date>