Derive a Time-to-Event Parameter

Add a time-to-event parameter to the input dataset.

derive_param_tte(
  dataset = NULL,
  dataset_adsl,
  source_datasets,
  by_vars = NULL,
  start_date = TRTSDT,
  start_date_imputation_flag = NULL,
  start_time_imputation_flag = NULL,
  event_conditions,
  censor_conditions,
  create_datetime = FALSE,
  set_values_to,
  subject_keys = vars(STUDYID, USUBJID)
)

Arguments

dataset
Input dataset

The PARAMCD variable is expected.

dataset_adsl
ADSL input dataset

The variables specified for start_date, start_imputation_flag, and subject_keys are expected.

source_datasets
Source datasets

A named list of datasets is expected. The dataset_name field of tte_source() refers to the dataset provided in the list.

by_vars
By variables

If the parameter is specified, separate time to event parameters are derived for each by group.

The by variables must be in at least one of the source datasets. Each source dataset must contain either all by variables or none of the by variables.

The by variables are not included in the output dataset.

start_date
Time to event origin date

The variable STARTDT is set to the specified date. The value is taken from the ADSL dataset.

If the event or censoring date is before the origin date, ADT is set to the origin date.

If the specified variable is imputed, the corresponding date imputation flag must specified for start_imputation_flag.

start_date_imputation_flag
Date imputation flag for start date

If the start date is imputed, the corresponding date imputation flag must be specified. The variable STARTDTF is set to the specified variable.

start_time_imputation_flag
Time imputation flag for start date

If the start time is imputed, the corresponding time imputation flag must be specified. The variable STARTTMF is set to the specified variable.

event_conditions
Sources and conditions defining events

A list of event_source() objects is expected.

censor_conditions
Sources and conditions defining censorings

A list of censor_source() objects is expected.

create_datetime
Create datetime variables?

If set to TRUE, variables ADTM and STARTDTM are created. Otherwise, variables ADT and STARTDT are created.

set_values_to
Variables to set

A named list returned by vars() defining the variables to be set for the new parameter, e.g. vars(PARAMCD = "OS", PARAM = "Overall Survival") is expected. The values must be symbols, character strings, numeric values, expressions, or NA.

subject_keys
Variables to uniquely identify a subject

A list of symbols created using vars() is expected.

Value

The input dataset with the new parameter added

Details

The following steps are performed to create the observations of the new parameter:

Deriving the events:

For each event source dataset the observations as specified by the filter element are selected. Then for each patient the first observation (with respect to date) is selected.

The ADT variable is set to the variable specified by the date element. If the date variable is a datetime variable, only the datepart is copied. If the source variable is a character variable, it is converted to a date. If the date is incomplete, it is imputed as the first possible date.

The CNSR variable is added and set to the censor element.

The variables specified by the set_values_to element are added.

The selected observations of all event source datasets are combined into a single dataset.

For each patient the first observation (with respect to the ADT variable) from the single dataset is selected.

Deriving the censoring observations:

For each censoring source dataset the observations as specified by the filter element are selected. Then for each patient the last observation (with respect to date) is selected.

The ADT variable is set to the variable specified by the date element. If the date variable is a datetime variable, only the datepart is copied. If the source variable is a character variable, it is converted to a date. If the date is incomplete, it is imputed as the first possible date.

The CNSR variable is added and set to the censor element.

The variables specified by the set_values_to element are added.

The selected observations of all censoring source datasets are combined into a single dataset.

For each patient the last observation (with respect to the ADT variable) from the single dataset is selected.

For each subject (as defined by the subject_keys parameter) an observation is selected. If an event is available, the event observation is selected. Otherwise the censoring observation is selected.

Finally

the variables specified for start_date and start_imputation_flag are joined from the ADSL dataset,

the variables as defined by the set_values_to parameter are added,

the ADT/ADTM variable is set to the maximum of ADT/ADTM and STARTDT/STARTDTM (depending on the create_datetime parameter), and

the new observations are added to the output dataset.

Examples

library(dplyr, warn.conflicts = FALSE)
library(lubridate, warn.conflicts = FALSE)
data("adsl")

death <- event_source(
  dataset_name = "adsl",
  filter = DTHFL == "Y",
  date = DTHDT,
  set_values_to = vars(
    EVNTDESC = "DEATH",
    SRCDOM = "ADSL",
    SRCVAR = "DTHDT"
  )
)

last_alive_dt <- censor_source(
  dataset_name = "adsl",
  date = LSTALVDT,
  set_values_to = vars(
    EVNTDESC = "LAST DATE KNOWN ALIVE",
    SRCDOM = "ADSL",
    SRCVAR = "LSTALVDT"
  )
)

derive_param_tte(
  dataset_adsl = adsl,
  event_conditions = list(death),
  censor_conditions = list(last_alive_dt),
  source_datasets = list(adsl = adsl),
  set_values_to = vars(
    PARAMCD = "OS",
    PARAM = "Overall Survival"
  )
) %>%
  select(-STUDYID) %>%
  filter(row_number() %in% 20:30)

#> # A tibble: 11 x 9
#>    USUBJID  EVNTDESC    SRCDOM SRCVAR  CNSR ADT        STARTDT    PARAMCD PARAM 
#>    <chr>    <chr>       <chr>  <chr>  <int> <date>     <date>     <chr>   <chr> 
#>  1 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-08-03 2013-02-02 OS      Overa…
#>  2 01-701-… DEATH       ADSL   DTHDT      0 2013-01-14 2012-11-15 OS      Overa…
#>  3 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-09-22 2013-03-30 OS      Overa…
#>  4 01-701-… LAST DATE … ADSL   LSTAL…     1 2014-07-11 2014-01-11 OS      Overa…
#>  5 01-701-… LAST DATE … ADSL   LSTAL…     1 2014-06-14 2014-02-07 OS      Overa…
#>  6 01-701-… LAST DATE … ADSL   LSTAL…     1 2014-07-26 2014-01-25 OS      Overa…
#>  7 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-06-14 2013-03-24 OS      Overa…
#>  8 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-11-05 2013-08-29 OS      Overa…
#>  9 01-701-… LAST DATE … ADSL   LSTAL…     1 2014-11-20 2014-05-22 OS      Overa…
#> 10 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-04-02 2012-10-02 OS      Overa…
#> 11 01-701-… LAST DATE … ADSL   LSTAL…     1 2013-02-07 2013-01-05 OS      Overa…

# derive time to adverse event for each preferred term #
adsl <- tibble::tribble(
  ~USUBJID, ~TRTSDT,           ~EOSDT,
  "01",     ymd("2020-12-06"), ymd("2021-03-06"),
  "02",     ymd("2021-01-16"), ymd("2021-02-03")
) %>%
  mutate(STUDYID = "AB42")

ae <- tibble::tribble(
  ~USUBJID, ~AESTDTC,           ~AESEQ, ~AEDECOD,
  "01",     "2021-01-03T10:56", 1,      "Flu",
  "01",     "2021-03-04",       2,      "Cough",
  "01",     "2021",             3,      "Flu"
) %>%
  mutate(STUDYID = "AB42")

ttae <- event_source(
  dataset_name = "ae",
  date = AESTDTC,
  set_values_to = vars(
    EVNTDESC = "AE",
    SRCDOM = "AE",
    SRCVAR = "AESTDTC",
    SRCSEQ = AESEQ
  )
)

eos <- censor_source(
  dataset_name = "adsl",
  date = EOSDT,
  set_values_to = vars(
    EVNTDESC = "END OF STUDY",
    SRCDOM = "ADSL",
    SRCVAR = "EOSDT"
  )
)

derive_param_tte(
  dataset_adsl = adsl,
  by_vars = vars(AEDECOD),
  start_date = TRTSDT,
  event_conditions = list(ttae),
  censor_conditions = list(eos),
  source_datasets = list(adsl = adsl, ae = ae),
  set_values_to = vars(
    PARAMCD = paste0("TTAE", as.numeric(as.factor(AEDECOD))),
    PARAM = paste("Time to First", AEDECOD, "Adverse Event"),
    PARCAT1 = "TTAE",
    PARCAT2 = AEDECOD
  )
) %>%
  select(USUBJID, STARTDT, PARAMCD, PARAM, ADT, CNSR, SRCSEQ)

#> # A tibble: 4 x 7
#>   USUBJID STARTDT    PARAMCD PARAM                       ADT         CNSR SRCSEQ
#>   <chr>   <date>     <chr>   <chr>                       <date>     <int>  <dbl>
#> 1 01      2020-12-06 TTAE1   Time to First Cough Advers… 2021-03-04     0      2
#> 2 01      2020-12-06 TTAE2   Time to First Flu Adverse … 2021-01-01     0      3
#> 3 02      2021-01-16 TTAE1   Time to First Cough Advers… 2021-02-03     1     NA
#> 4 02      2021-01-16 TTAE2   Time to First Flu Adverse … 2021-02-03     1     NA