Add New Records Within By Groups Using Aggregation Functions

It is not uncommon to have an analysis need whereby one needs to derive an analysis value (AVAL) from multiple records. The ADaM basic dataset structure variable DTYPE is available to indicate when a new derived records has been added to a dataset.

derive_summary_records(
  dataset,
  by_vars,
  fns,
  filter_rows = NULL,
  set_values_to = NULL,
  drop_values_from = NULL
)

Arguments

dataset
A data frame.

by_vars
Variables to consider for generation of groupwise summary records. Providing the names of variables in vars() will create a groupwise summary and generate summary records for the specified groups.

fns
List of formulas specifying variable to use for aggregations. This can include base functions like mean(), min(), max(), median(), sd(), or sum() or any other user-defined aggregation function. For example,

When a summary function is same for one or more analysis variable, use fns = list(vars(AVAL, CHG) ~ mean).

If a different summary function is required for each analysis variable, use fns = list(AVAL ~ mean, CHG ~ sum(., na.rm = TRUE)).

In general,

LHS refer to the one or more variable to use for summarizing.

RHS refer to a single summary function.

In the formula representation e.g., CHG ~ sum(., na.rm = TRUE), a . serves as the data to be summarized which refers to the variable CHG.

filter_rows
Filter condition as logical expression to apply during summary calculation. By default, filtering expressions are computed within by_vars as this will help when an aggregating, lagging, or ranking function is involved.

For example,

filter_rows = (AVAL > mean(AVAL, na.rm = TRUE)) will filter all AVAL values greater than mean of AVAL with in by_vars.

filter_rows = (dplyr::n() > 2) will filter n count of by_vars greater than 2.

set_values_to
A list of variable name-value pairs. Use this argument if you need to change the values of any newly derived records. Always new values in set_values_to should be equal to the length of analysis variable used in the fns argument. For example,

On using single analysis variable in fns = list(AVAL ~ mean) and set_values_to = vars(AVISITN = 9999, AVISIT = "Endpoint") would change the value of AVISITN to 9999 and AVISIT to Endpoint instead of retaining.

Multiple analysis variables in fns = list(vars(AVAL, CHG) ~ mean) and set_values_to = vars(AVISITN = c(9998, 9999)) would change AVISITN to 9998 for AVAL and AVISITN to 9999 for CHG.

drop_values_from
Providing the names of variables in vars() will drop values and set as missing.

Value

A data frame with derived records appended to original dataset.

Details

When all records have same values within by_vars then this function will retain those common values in the newly derived records. Otherwise new value will be set to NA.

Examples

adeg <- tibble::tribble(
  ~USUBJID, ~EGSEQ, ~PARAM,             ~AVISIT,    ~EGDTC,            ~AVAL, ~TRTA,
  "XYZ-1001",    1, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:50",  385, "",
  "XYZ-1001",    2, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:52",  399, "",
  "XYZ-1001",    3, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:56",  396, "",
  "XYZ-1001",    4, "QTcF Int. (msec)", "Visit 2",  "2016-03-08T09:45",  384, "Placebo",
  "XYZ-1001",    5, "QTcF Int. (msec)", "Visit 2",  "2016-03-08T09:48",  393, "Placebo",
  "XYZ-1001",    6, "QTcF Int. (msec)", "Visit 2",  "2016-03-08T09:51",  388, "Placebo",
  "XYZ-1001",    7, "QTcF Int. (msec)", "Visit 3",  "2016-03-22T10:45",  385, "Placebo",
  "XYZ-1001",    8, "QTcF Int. (msec)", "Visit 3",  "2016-03-22T10:48",  394, "Placebo",
  "XYZ-1001",    9, "QTcF Int. (msec)", "Visit 3",  "2016-03-22T10:51",  402, "Placebo",
  "XYZ-1002",    1, "QTcF Int. (msec)", "Baseline", "2016-02-22T07:58",  399, "",
  "XYZ-1002",    2, "QTcF Int. (msec)", "Baseline", "2016-02-22T07:58",  410, "",
  "XYZ-1002",    3, "QTcF Int. (msec)", "Baseline", "2016-02-22T08:01",  392, "",
  "XYZ-1002",    4, "QTcF Int. (msec)", "Visit 2",  "2016-03-06T09:50",  401, "Active 20mg",
  "XYZ-1002",    5, "QTcF Int. (msec)", "Visit 2",  "2016-03-06T09:53",  407, "Active 20mg",
  "XYZ-1002",    6, "QTcF Int. (msec)", "Visit 2",  "2016-03-06T09:56",  400, "Active 20mg",
  "XYZ-1002",    7, "QTcF Int. (msec)", "Visit 3",  "2016-03-24T10:50",  412, "Active 20mg",
  "XYZ-1002",    8, "QTcF Int. (msec)", "Visit 3",  "2016-03-24T10:53",  414, "Active 20mg",
  "XYZ-1002",    9, "QTcF Int. (msec)", "Visit 3",  "2016-03-24T10:56",  402, "Active 20mg",
)

# Summarize the average of the triplicate ECG interval values (AVAL)
derive_summary_records(
  adeg,
  by_vars = vars(USUBJID, PARAM, AVISIT),
  fns = list(AVAL ~ mean(., na.rm = TRUE)),
  set_values_to = vars(DTYPE = "AVERAGE")
)

#> # A tibble: 24 x 8
#>    USUBJID  EGSEQ PARAM           AVISIT   EGDTC            AVAL TRTA     DTYPE 
#>    <chr>    <dbl> <chr>           <chr>    <chr>           <dbl> <chr>    <chr> 
#>  1 XYZ-1001     1 QTcF Int. (mse… Baseline 2016-02-24T07:…  385  ""       <NA>  
#>  2 XYZ-1001     2 QTcF Int. (mse… Baseline 2016-02-24T07:…  399  ""       <NA>  
#>  3 XYZ-1001     3 QTcF Int. (mse… Baseline 2016-02-24T07:…  396  ""       <NA>  
#>  4 XYZ-1001    NA QTcF Int. (mse… Baseline <NA>             393. ""       AVERA…
#>  5 XYZ-1001     4 QTcF Int. (mse… Visit 2  2016-03-08T09:…  384  "Placeb… <NA>  
#>  6 XYZ-1001     5 QTcF Int. (mse… Visit 2  2016-03-08T09:…  393  "Placeb… <NA>  
#>  7 XYZ-1001     6 QTcF Int. (mse… Visit 2  2016-03-08T09:…  388  "Placeb… <NA>  
#>  8 XYZ-1001    NA QTcF Int. (mse… Visit 2  <NA>             388. "Placeb… AVERA…
#>  9 XYZ-1001     7 QTcF Int. (mse… Visit 3  2016-03-22T10:…  385  "Placeb… <NA>  
#> 10 XYZ-1001     8 QTcF Int. (mse… Visit 3  2016-03-22T10:…  394  "Placeb… <NA>  
#> # … with 14 more rows

advs <- tibble::tribble(
  ~USUBJID,     ~VSSEQ, ~PARAM,  ~AVAL, ~VSSTRESU, ~VISIT,      ~VSDTC,
  "XYZ-001-001",  1164, "Weight",   99, "kg",      "Screening", "2018-03-19",
  "XYZ-001-001",  1165, "Weight",  101, "kg",      "Run-In"   , "2018-03-26",
  "XYZ-001-001",  1166, "Weight",  100, "kg",      "Baseline" , "2018-04-16",
  "XYZ-001-001",  1167, "Weight",   94, "kg",      "Week 24"  , "2018-09-30",
  "XYZ-001-001",  1168, "Weight",   92, "kg",      "Week 48"  , "2019-03-17",
  "XYZ-001-001",  1169, "Weight",   95, "kg",      "Week 52"  , "2019-04-14",
)

# Set new values to any variable. Here, `DTYPE = MAXIMUM` refers to `max()` records
# and `DTYPE = AVERAGE` refers to `mean()` records.
# `set_values_to` must be of the same length as `fns`
derive_summary_records(
  advs,
  by_vars = vars(USUBJID, PARAM),
  fns = list(AVAL ~ max, AVAL ~ mean),
  set_values_to = vars(DTYPE = c("MAXIMUM", "AVERAGE"))
)

#> # A tibble: 8 x 8
#>   USUBJID     VSSEQ PARAM   AVAL VSSTRESU VISIT     VSDTC      DTYPE  
#>   <chr>       <dbl> <chr>  <dbl> <chr>    <chr>     <chr>      <chr>  
#> 1 XYZ-001-001  1164 Weight  99   kg       Screening 2018-03-19 <NA>   
#> 2 XYZ-001-001  1165 Weight 101   kg       Run-In    2018-03-26 <NA>   
#> 3 XYZ-001-001  1166 Weight 100   kg       Baseline  2018-04-16 <NA>   
#> 4 XYZ-001-001  1167 Weight  94   kg       Week 24   2018-09-30 <NA>   
#> 5 XYZ-001-001  1168 Weight  92   kg       Week 48   2019-03-17 <NA>   
#> 6 XYZ-001-001  1169 Weight  95   kg       Week 52   2019-04-14 <NA>   
#> 7 XYZ-001-001    NA Weight 101   kg       <NA>      <NA>       MAXIMUM
#> 8 XYZ-001-001    NA Weight  96.8 kg       <NA>      <NA>       AVERAGE

# Drop retained value of VSSTRESU in the derived record
derive_summary_records(
  advs,
  by_vars = vars(USUBJID, PARAM),
  fns = list(AVAL ~ mean),
  set_values_to = vars(DTYPE = "MAXIMUM"),
  drop_values_from = vars(VSSTRESU)
)

#> # A tibble: 7 x 8
#>   USUBJID     VSSEQ PARAM   AVAL VSSTRESU VISIT     VSDTC      DTYPE  
#>   <chr>       <dbl> <chr>  <dbl> <chr>    <chr>     <chr>      <chr>  
#> 1 XYZ-001-001  1164 Weight  99   kg       Screening 2018-03-19 <NA>   
#> 2 XYZ-001-001  1165 Weight 101   kg       Run-In    2018-03-26 <NA>   
#> 3 XYZ-001-001  1166 Weight 100   kg       Baseline  2018-04-16 <NA>   
#> 4 XYZ-001-001  1167 Weight  94   kg       Week 24   2018-09-30 <NA>   
#> 5 XYZ-001-001  1168 Weight  92   kg       Week 48   2019-03-17 <NA>   
#> 6 XYZ-001-001  1169 Weight  95   kg       Week 52   2019-04-14 <NA>   
#> 7 XYZ-001-001    NA Weight  96.8 <NA>     <NA>      <NA>       MAXIMUM

# Sample ADEG dataset with triplicate record for only AVISIT = 'Baseline' ---
adeg <- tibble::tribble(
  ~USUBJID, ~EGSEQ, ~PARAM,             ~AVISIT,    ~EGDTC,            ~AVAL, ~TRTA,
  "XYZ-1001",    1, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:50",  385, "",
  "XYZ-1001",    2, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:52",  399, "",
  "XYZ-1001",    3, "QTcF Int. (msec)", "Baseline", "2016-02-24T07:56",  396, "",
  "XYZ-1001",    4, "QTcF Int. (msec)", "Visit 2",  "2016-03-08T09:48",  393, "Placebo",
  "XYZ-1001",    5, "QTcF Int. (msec)", "Visit 2",  "2016-03-08T09:51",  388, "Placebo",
  "XYZ-1001",    6, "QTcF Int. (msec)", "Visit 3",  "2016-03-22T10:48",  394, "Placebo",
  "XYZ-1001",    7, "QTcF Int. (msec)", "Visit 3",  "2016-03-22T10:51",  402, "Placebo",
  "XYZ-1002",    1, "QTcF Int. (msec)", "Baseline", "2016-02-22T07:58",  399, "",
  "XYZ-1002",    2, "QTcF Int. (msec)", "Baseline", "2016-02-22T07:58",  410, "",
  "XYZ-1002",    3, "QTcF Int. (msec)", "Baseline", "2016-02-22T08:01",  392, "",
  "XYZ-1002",    4, "QTcF Int. (msec)", "Visit 2",  "2016-03-06T09:53",  407, "Active 20mg",
  "XYZ-1002",    5, "QTcF Int. (msec)", "Visit 2",  "2016-03-06T09:56",  400, "Active 20mg",
  "XYZ-1002",    6, "QTcF Int. (msec)", "Visit 3",  "2016-03-24T10:53",  414, "Active 20mg",
  "XYZ-1002",    7, "QTcF Int. (msec)", "Visit 3",  "2016-03-24T10:56",  402, "Active 20mg",
)

# Summarize the average of AVAL for AVISIT records greater than 2
derive_summary_records(
  adeg,
  by_vars = vars(USUBJID, PARAM, AVISIT),
  fns = list(AVAL ~ mean(., na.rm = TRUE)),
  filter_rows = dplyr::n() > 2,
  set_values_to = vars(DTYPE = "AVERAGE")
)

#> # A tibble: 16 x 8
#>    USUBJID  EGSEQ PARAM          AVISIT   EGDTC          AVAL TRTA        DTYPE 
#>    <chr>    <dbl> <chr>          <chr>    <chr>         <dbl> <chr>       <chr> 
#>  1 XYZ-1001     1 QTcF Int. (ms… Baseline 2016-02-24T0…  385  ""          <NA>  
#>  2 XYZ-1001     2 QTcF Int. (ms… Baseline 2016-02-24T0…  399  ""          <NA>  
#>  3 XYZ-1001     3 QTcF Int. (ms… Baseline 2016-02-24T0…  396  ""          <NA>  
#>  4 XYZ-1001    NA QTcF Int. (ms… Baseline <NA>           393. ""          AVERA…
#>  5 XYZ-1001     4 QTcF Int. (ms… Visit 2  2016-03-08T0…  393  "Placebo"   <NA>  
#>  6 XYZ-1001     5 QTcF Int. (ms… Visit 2  2016-03-08T0…  388  "Placebo"   <NA>  
#>  7 XYZ-1001     6 QTcF Int. (ms… Visit 3  2016-03-22T1…  394  "Placebo"   <NA>  
#>  8 XYZ-1001     7 QTcF Int. (ms… Visit 3  2016-03-22T1…  402  "Placebo"   <NA>  
#>  9 XYZ-1002     1 QTcF Int. (ms… Baseline 2016-02-22T0…  399  ""          <NA>  
#> 10 XYZ-1002     2 QTcF Int. (ms… Baseline 2016-02-22T0…  410  ""          <NA>  
#> 11 XYZ-1002     3 QTcF Int. (ms… Baseline 2016-02-22T0…  392  ""          <NA>  
#> 12 XYZ-1002    NA QTcF Int. (ms… Baseline <NA>           400. ""          AVERA…
#> 13 XYZ-1002     4 QTcF Int. (ms… Visit 2  2016-03-06T0…  407  "Active 20… <NA>  
#> 14 XYZ-1002     5 QTcF Int. (ms… Visit 2  2016-03-06T0…  400  "Active 20… <NA>  
#> 15 XYZ-1002     6 QTcF Int. (ms… Visit 3  2016-03-24T1…  414  "Active 20… <NA>  
#> 16 XYZ-1002     7 QTcF Int. (ms… Visit 3  2016-03-24T1…  402  "Active 20… <NA>