Derive BASETYPE Variable

Adds the BASETYPE variable to a dataset and duplicates records based upon the provided conditions

derive_var_basetype(dataset, basetypes)

Arguments

dataset
Input dataset

The columns specified in the expressions inside basetypes are required.

basetypes
A named list of expressions created using exprs()

The names corresponds to the values of the newly created BASETYPE variables and the expressions are used to subset the input dataset.

Details

For each element of basetypes the input dataset is subset based upon the provided expression and the BASETYPE variable is set to the name of the expression. Then, all subsets are stacked. Records which do not match any condition are kept and BASETYPE is set to NA.

Examples

bds <- tibble::tribble(
  ~USUBJID, ~EPOCH,         ~PARAMCD,  ~ASEQ, ~AVAL,
  "P01",    "RUN-IN",       "PARAM01", 1,     10,
  "P01",    "RUN-IN",       "PARAM01", 2,      9.8,
  "P01",    "DOUBLE-BLIND", "PARAM01", 3,      9.2,
  "P01",    "DOUBLE-BLIND", "PARAM01", 4,     10.1,
  "P01",    "OPEN-LABEL",   "PARAM01", 5,     10.4,
  "P01",    "OPEN-LABEL",   "PARAM01", 6,      9.9,
  "P02",    "RUN-IN",       "PARAM01", 1,     12.1,
  "P02",    "DOUBLE-BLIND", "PARAM01", 2,     10.2,
  "P02",    "DOUBLE-BLIND", "PARAM01", 3,     10.8,
  "P02",    "OPEN-LABEL",   "PARAM01", 4,     11.4,
  "P02",    "OPEN-LABEL",   "PARAM01", 5,     10.8
)

derive_var_basetype(
  dataset = bds,
  basetypes = exprs(
    "RUN-IN" = EPOCH %in% c("RUN-IN", "STABILIZATION", "DOUBLE-BLIND", "OPEN-LABEL"),
    "DOUBLE-BLIND" = EPOCH %in% c("DOUBLE-BLIND", "OPEN-LABEL"),
    "OPEN-LABEL" = EPOCH == "OPEN-LABEL"
  )
)

#> # A tibble: 23 x 6
#>    USUBJID EPOCH        PARAMCD  ASEQ  AVAL BASETYPE
#>    <chr>   <chr>        <chr>   <dbl> <dbl> <chr>   
#>  1 P01     RUN-IN       PARAM01     1  10   RUN-IN  
#>  2 P01     RUN-IN       PARAM01     2   9.8 RUN-IN  
#>  3 P01     DOUBLE-BLIND PARAM01     3   9.2 RUN-IN  
#>  4 P01     DOUBLE-BLIND PARAM01     4  10.1 RUN-IN  
#>  5 P01     OPEN-LABEL   PARAM01     5  10.4 RUN-IN  
#>  6 P01     OPEN-LABEL   PARAM01     6   9.9 RUN-IN  
#>  7 P02     RUN-IN       PARAM01     1  12.1 RUN-IN  
#>  8 P02     DOUBLE-BLIND PARAM01     2  10.2 RUN-IN  
#>  9 P02     DOUBLE-BLIND PARAM01     3  10.8 RUN-IN  
#> 10 P02     OPEN-LABEL   PARAM01     4  11.4 RUN-IN  
#> # … with 13 more rows