Filter out the first or last observation for each group

Filters out the first or last observation for each group.

filter_extreme(dataset, order, by_vars, mode = "last",
  check_type = "warning")

Arguments

dataset
Input dataset

The variables specified by the order and the by_vars parameter are expected.

order
Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables

by_vars
Grouping variables

Permitted Values: list of variables

mode
Selection mode (first or last)

If "first" is specified, the first observation of each by group is included in the output dataset. If "last" is specified, the last observation of each by group is included in the output dataset.

Permitted Values: "first", "last"

check_type
Check uniqueness?

If "warning" or "error" is specified, the specified message is issued if the observations of the input dataset are not unique with respect to the by variables and the order.

Default: "none"

Permitted Values: "none", "warning", "error"

Value

A dataset containing the first or last observation of each by group

Details

For each group (with respect to the variables specified for the by_vars parameter) the first or last observation (with respect to the order specified for the order parameter and the mode specified for the mode parameter) is included in the output dataset.

Examples

library(dplyr)
library(magrittr)

data("ex")

# selecting first dose for each patient
filter_extreme(ex,
               order = exprs(EXSEQ),
               by_vars = exprs(USUBJID),
               mode = 'first') %>%
  select(USUBJID, EXSEQ)

#> # A tibble: 254 x 2
#>    USUBJID     EXSEQ
#>    <chr>       <dbl>
#>  1 01-701-1015     1
#>  2 01-701-1023     1
#>  3 01-701-1028     1
#>  4 01-701-1033     1
#>  5 01-701-1034     1
#>  6 01-701-1047     1
#>  7 01-701-1097     1
#>  8 01-701-1111     1
#>  9 01-701-1115     1
#> 10 01-701-1118     1
#> # … with 244 more rows

# selecting highest dose for each patient
filter_extreme(ex,
               order = exprs(EXDOSE),
               by_vars = exprs(USUBJID),
               check_type = "none") %>%
  select(USUBJID, EXDOSE)

#> # A tibble: 254 x 2
#>    USUBJID     EXDOSE
#>    <chr>        <dbl>
#>  1 01-701-1015      0
#>  2 01-701-1023      0
#>  3 01-701-1028     81
#>  4 01-701-1033     54
#>  5 01-701-1034     81
#>  6 01-701-1047      0
#>  7 01-701-1097     54
#>  8 01-701-1111     54
#>  9 01-701-1115     54
#> 10 01-701-1118      0
#> # … with 244 more rows