Create an query object

A query object defines a query, e.g., a Standard MedDRA Query (SMQ), a Standardised Drug Grouping (SDG), or a customized query (CQ). It is used as input to create_query_data().

Usage

query(prefix, name = auto, id = NULL, add_scope_num = FALSE, definition = NULL)

Arguments

prefix

The value is used to populate VAR_PREFIX in the output dataset of create_query_data(), e.g., "SMQ03"

name

The value is used to populate QUERY_NAME in the output dataset of create_query_data(). If the auto keyword is specified, the variable is set to the name of the query in the SMQ/SDG database.

Permitted Values: A character scalar or the auto keyword. The auto keyword is permitted only for queries which are defined by an smq_select() or sdg_select() object.

id

The value is used to populate QUERY_ID in the output dataset of create_query_data(). If the auto keyword is specified, the variable is set to the id of the query in the SMQ/SDG database.

Permitted Values: A integer scalar or the auto keyword. The auto keyword is permitted only for queries which are defined by an smq_select() or sdg_select() object.

add_scope_num

Determines if QUERY_SCOPE_NUM in the output dataset of create_query_data() is populated

If the parameter is set to TRUE, the definition must be an smq_select() object.

Default: FALSE

Permitted Values: TRUE, FALSE

definition

Definition of terms belonging to the query

There are four different ways to define the terms:

An smq_select() object is specified to select a query from the SMQ database.

An sdg_select() object is specified to select a query from the SDG database.

A data frame with columns TERM_LEVEL and TERM_NAME or TERM_ID can be specified to define the terms of a customized query. The TERM_LEVEL should be set to the name of the variable which should be used to select the terms, e.g., "AEDECOD" or "AELLTCD". TERM_LEVEL does not need to be constant within a query. For example a query can be based on AEDECOD and AELLT.

If TERM_LEVEL refers to a character variable, TERM_NAME should be set to the value the variable. If it refers to a numeric variable, TERM_ID should be set to the value of the variable. If only character variables or only numeric variables are used, TERM_ID or TERM_NAME respectively can be omitted.

A list of data frames and smq_select() objects can be specified to define a customized query based on custom terms and SMQs. The data frames must have the same structure as described for the previous item.

Permitted Values: an smq_select() object, an sdg_select() object, a data frame, or a list of data frames and smq_select() objects.

Value

An object of class query.

See also

Author

Stefan Bundfuss

Examples

# create a query for an SMQ
library(tibble)
library(magrittr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)

query(
  prefix = "SMQ02",
  id = auto,
  definition = smq_select(
    name = "Pregnancy and neonatal topics (SMQ)",
    scope = "NARROW"
  )
)
#> $prefix
#> [1] "SMQ02"
#> 
#> $name
#> <quosure>
#> expr: ^auto
#> env:  0x5643d26a58d8
#> 
#> $id
#> <quosure>
#> expr: ^auto
#> env:  0x5643cbe63f10
#> 
#> $add_scope_num
#> [1] FALSE
#> 
#> $definition
#> $name
#> [1] "Pregnancy and neonatal topics (SMQ)"
#> 
#> $id
#> NULL
#> 
#> $scope
#> [1] "NARROW"
#> 
#> attr(,"class")
#> [1] "smq_select" "list"      
#> 
#> attr(,"class")
#> [1] "query" "list" 

# create a query for an SDG
query(
  prefix = "SDG01",
  id = auto,
  definition = sdg_select(
    name = "5-aminosalicylates for ulcerative colitis"
  )
)
#> $prefix
#> [1] "SDG01"
#> 
#> $name
#> <quosure>
#> expr: ^auto
#> env:  0x5643cbab2c48
#> 
#> $id
#> <quosure>
#> expr: ^auto
#> env:  0x5643cbe63f10
#> 
#> $add_scope_num
#> [1] FALSE
#> 
#> $definition
#> $name
#> [1] "5-aminosalicylates for ulcerative colitis"
#> 
#> $id
#> NULL
#> 
#> attr(,"class")
#> [1] "sdg_select" "list"      
#> 
#> attr(,"class")
#> [1] "query" "list" 

# creating a query for a customized query
cqterms <- tribble(
  ~TERM_NAME, ~TERM_ID,
  "APPLICATION SITE ERYTHEMA", 10003041L,
  "APPLICATION SITE PRURITUS", 10003053L
) %>%
  mutate(TERM_LEVEL = "AEDECOD")

query(
  prefix = "CQ01",
  name = "Application Site Issues",
  definition = cqterms
)
#> $prefix
#> [1] "CQ01"
#> 
#> $name
#> [1] "Application Site Issues"
#> 
#> $add_scope_num
#> [1] FALSE
#> 
#> $definition
#> # A tibble: 2 x 3
#>   TERM_NAME                  TERM_ID TERM_LEVEL
#>   <chr>                        <int> <chr>     
#> 1 APPLICATION SITE ERYTHEMA 10003041 AEDECOD   
#> 2 APPLICATION SITE PRURITUS 10003053 AEDECOD   
#> 
#> attr(,"class")
#> [1] "query" "list" 

# creating a customized query based on SMQs and additional terms
query(
  prefix = "CQ03",
  name = "Special issues of interest",
  definition = list(
    cqterms,
    smq_select(
      name = "Pregnancy and neonatal topics (SMQ)",
      scope = "NARROW"
    ),
    smq_select(
      id = 8050L,
      scope = "BROAD"
    )
  )
)
#> $prefix
#> [1] "CQ03"
#> 
#> $name
#> [1] "Special issues of interest"
#> 
#> $add_scope_num
#> [1] FALSE
#> 
#> $definition
#> $definition[[1]]
#> # A tibble: 2 x 3
#>   TERM_NAME                  TERM_ID TERM_LEVEL
#>   <chr>                        <int> <chr>     
#> 1 APPLICATION SITE ERYTHEMA 10003041 AEDECOD   
#> 2 APPLICATION SITE PRURITUS 10003053 AEDECOD   
#> 
#> $definition[[2]]
#> $name
#> [1] "Pregnancy and neonatal topics (SMQ)"
#> 
#> $id
#> NULL
#> 
#> $scope
#> [1] "NARROW"
#> 
#> attr(,"class")
#> [1] "smq_select" "list"      
#> 
#> $definition[[3]]
#> $name
#> NULL
#> 
#> $id
#> [1] 8050
#> 
#> $scope
#> [1] "BROAD"
#> 
#> attr(,"class")
#> [1] "smq_select" "list"      
#> 
#> 
#> attr(,"class")
#> [1] "query" "list"