Create a tte_source Object

The tte_source object is used to define events and possible censorings.

tte_source(dataset_name, filter = NULL, date, censor = 0, set_values_to = NULL)

Arguments

dataset_name
The name of the source dataset

The name refers to the dataset provided by the source_datasets parameter of derive_param_tte().

filter
An unquoted condition for selecting the observations from dataset which are events or possible censoring time points.

date
A variable providing the date of the event or censoring. A date, a datetime, or a character variable containing ISO 8601 dates can be specified. An unquoted symbol is expected.

censor
Censoring value

CDISC strongly recommends using 0 for events and positive integers for censoring.

set_values_to
A named list returned by vars() defining the variables to be set for the event or censoring, e.g. vars(EVENTDESC = "DEATH", SRCDOM = "ADSL", SRCVAR = "DTHDT"). The values must be a symbol, a character string, a numeric value, or NA.

Value

An object of class tte_source

See also

Examples

# Death event
admiral:::tte_source(
  dataset_name = "adsl",
  filter = DTHFL == "Y",
  date = DTHDT,
  censor = 0,
  set_values_to = vars(
    EVNTDESC = "DEATH",
    SRCDOM = "ADSL",
    SRCVAR = "DTHDT"
  )
)

#> <tte_source> object
#> dataset_name: "adsl"
#> filter: DTHFL == "Y"
#> date: DTHDT
#> censor: 0
#> set_values_to:
#>   EVNTDESC: "DEATH"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "DTHDT"

# Last study date known alive censor
admiral:::tte_source(
  dataset_name = "adsl",
  date = LSTALVDT,
  censor = 1,
  set_values_to = vars(
    EVNTDESC = "ALIVE",
    SRCDOM = "ADSL",
    SRCVAR = "LSTALVDT"
  )
)

#> <tte_source> object
#> dataset_name: "adsl"
#> filter: NULL
#> date: LSTALVDT
#> censor: 1
#> set_values_to:
#>   EVNTDESC: "ALIVE"
#>   SRCDOM: "ADSL"
#>   SRCVAR: "LSTALVDT"